% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_aggregationTree_information.R,
%   R/add_aggregationTree_information_toRxsTree.R
\name{add_aggregationTree_information}
\alias{add_aggregationTree_information}
\alias{add_aggregationTree_information_toRxsTree}
\title{Add information from an aggregation tree to a Rxs tree}
\usage{
add_aggregationTree_information(
  x,
  aggregationTree,
  fieldName,
  prefixes = NULL,
  suffixes = NULL
)

add_aggregationTree_information_toRxsTree(
  rxsTree,
  aggregationTree,
  fieldName,
  prefixes = NULL,
  suffixes = NULL
)
}
\arguments{
\item{aggregationTree}{The aggregation tree}

\item{fieldName}{The field storing the values along which to aggregate}

\item{prefixes, suffixes}{Vector of pre- and suffixes to add to the
\code{fieldname} when storing the values for progressive aggregation}

\item{rxsTree, x}{The Rxs tree (as \code{rxsTree}) or the object with
multiple Rxs trees (as \code{x})}
}
\value{
Invisibly (and, irrelevantly), the Rxs tree
}
\description{
Add information from an aggregation tree to a Rxs tree
}
