% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibentry_to_df.R
\name{as.data.frame.person}
\alias{as.data.frame.person}
\title{Convert \code{\link[utils:person]{utils::person()}} object to a data frame}
\usage{
\method{as.data.frame}{person}(x, ...)
}
\arguments{
\item{x}{The \code{\link[utils:person]{utils::person()}} object}

\item{...}{Any additional argument to pass to \code{\link[base:data.frame]{base::data.frame()}}.}
}
\value{
A data frame with one row for each person.
}
\description{
This function takes a \code{\link[utils:person]{utils::person()}} object and converts it to a data frame.
}
