% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibentry_to_df.R
\name{bibentry_to_df}
\alias{bibentry_to_df}
\title{Convert a bibentry (or more) to a data frame}
\usage{
bibentry_to_df(bibentry, authorCollapseFunc = metabefor::vecTxt)
}
\arguments{
\item{bibentry}{The bibentry (or bibentries)}

\item{authorCollapseFunc}{The function to use to collapse the vector of
author names into one value.}
}
\value{
A data frame
}
\description{
This function takes a bibentry and converts it to a data frame.
}
\examples{
### Use the base R bibentry
metabefor::bibentry_to_df(citation());
}
