% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate_sources.R
\name{check_duplicate_sources}
\alias{check_duplicate_sources}
\title{Check for duplicate sources}
\usage{
check_duplicate_sources(
  primarySources,
  secondarySources = NULL,
  useStringDistances = FALSE,
  stringDistance = 5,
  stringDistanceMethod = "osa",
  charsToZap = "[^A-Za-z0-9]",
  doiCol = "doi",
  matchFully = c("year", "title", "author"),
  matchStart = c(title = 20, author = 10),
  matchEnd = c(title = 20, author = 10),
  forDeduplicationSuffix = "_forDeduplication",
  returnRawStringDistances = FALSE,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{primarySources}{The primary dataframe with sources}

\item{secondarySources}{The secondary dataframe with sources}

\item{useStringDistances}{Whether to use string distances - note that
that can be very slow and take along time if you have thousands of sources.}

\item{stringDistance}{The string distance for titles}

\item{stringDistanceMethod}{Method to use for string distance computation}

\item{charsToZap}{The characters to delete from fields before looking for duplicates}

\item{doiCol}{The name of the column with the DOIs}

\item{forDeduplicationSuffix}{Suffix to add to optional deduplication columns}

\item{returnRawStringDistances}{Whether to return the raw string distances
or not (this can be \emph{very} large).}

\item{silent}{Whether to be silent or chatty.}
}
\value{
A vector indicating for each record whether it's a duplicate, with
an attribute called \code{duplicateInfo} that holds more detailed information
and that can be accessed using the \code{\link[=attributes]{attributes()}} function.
}
\description{
Check for duplicate sources
}
