% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneRxsProject.R
\name{cloneRxsProject}
\alias{cloneRxsProject}
\title{Clone a full Rxs project object}
\usage{
cloneRxsProject(x, silent = metabefor::opts$get("silent"))
}
\arguments{
\item{x}{The full Rxs project object}

\item{silent}{Whether to be chatty or silent}
}
\value{
The cloned object
}
\description{
Clone a full Rxs project object
}
\examples{
### Load an example Rxs project
example_rxsProject <-
  metabefor::example_rxsProject;
  
### Check whether in this object, value 'hi' exists:
example_rxsProject$rxsTrees$qurid_7h50rzpq$hi;

### (should be NULL)

### Try to make a copy 'the conventional way'
copy_rxsProject <-
  example_rxsProject;
  
### Change something
copy_rxsProject$rxsTrees$qurid_7h50rzpq$hi <- "Hello";

### Check that worked
copy_rxsProject$rxsTrees$qurid_7h50rzpq$hi;

### Show how magically stuff also changed in the
### original even though we didn't change anything
example_rxsProject$rxsTrees$qurid_7h50rzpq$hi;

### Clone it instead
copy_rxsProject <-
  metabefor::cloneRxsProject(example_rxsProject);
  
### Make thes ame change in the copy
copy_rxsProject$rxsTrees$qurid_7h50rzpq$hi <- "Goodbye";

### Show how the value in the original didn't change again
example_rxsProject$rxsTrees$qurid_7h50rzpq$hi;
}
