% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneStudiesObject.R
\name{cloneStudiesObject}
\alias{cloneStudiesObject}
\title{Clone a parsed Rxs object (a 'studies object')}
\usage{
cloneStudiesObject(x, silent = metabefor::opts$get("silent"))
}
\arguments{
\item{x}{The studies object}

\item{silent}{Whether to be chatty or silent}
}
\value{
The cloned object
}
\description{
Clone a parsed Rxs object (a 'studies object')
}
