% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrycodelist.R
\name{countrycodelist}
\alias{countrycodelist}
\title{Return a list of country codes}
\usage{
countrycodelist(type = "alpha1")
}
\arguments{
\item{type}{The type; currently only \code{"alpha1"} is a valid value.}
}
\value{
A character vector.
}
\description{
Return a list of country codes
}
\examples{
countrycodelist()
}
