% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_get_stringDistances.R
\name{dedup_get_stringDistances}
\alias{dedup_get_stringDistances}
\title{Get string distances, optionally using multiple cores}
\usage{
dedup_get_stringDistances(
  x,
  y = x,
  stringDistanceMethod = "osa",
  parallel = FALSE,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{x}{The first vector}

\item{y}{The second vector}

\item{stringDistanceMethod}{The method to use (\code{"osa"}, \code{"lv"}, \code{"dl"},
\code{"hamming"}, \code{"lcs"}, \code{"qgram"}, \code{"cosine"}, \code{"jaccard"}, \code{"jw"}, or
\code{"soundex"})}

\item{parallel}{Whether to use parallel processing}

\item{silent}{Whether to be silent or chatty}
}
\value{
A data frame with the string distances between the elements
}
\description{
Get string distances, optionally using multiple cores
}
\examples{
### Create two vectors with strings
a <- c("Apple", "Blueberry", "Cherry", "Date");
c <- c("Airplane", "Bus", "Cycle");

dedup_get_stringDistances(
  a,
  b
);
}
