% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_matches_trimmed.R
\name{dedup_matches_trimmed}
\alias{dedup_matches_trimmed}
\title{Return for needles whether they occur in the haystack, trimmed if desired}
\usage{
dedup_matches_trimmed(needles, haystack, start = 0, end = 0)
}
\arguments{
\item{needles}{The entries to look for}

\item{haystack}{The set of entries in which to look}

\item{start}{The number of characters to use from the beginning (truncating
after this number of characters, e.g. only looking at the first 20 characters)}

\item{end}{The number of characters to use from the end (trimming everything
before this number of characters, e.g. only looking at the last 10 characters)}
}
\value{
A logical vector
}
\description{
Return for needles whether they occur in the haystack, trimmed if desired
}
\examples{
dedup_matches_trimmed(
  c('a', 'b', 'c'),
  c('c', 'd');
);
}
