% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultGraphTheme.R
\name{defaultGraphTheme}
\alias{defaultGraphTheme}
\title{Get default graph theme from options}
\usage{
defaultGraphTheme()
}
\value{
\code{metabefor::opts$get('defaultGraphTheme')}
}
\description{
Convenience function that just returns
\code{metabefor::opts$get('defaultGraphTheme')}.
}
