% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deinterlace_file.R
\name{deinterlace_file}
\alias{deinterlace_file}
\title{Deinterlace a file}
\usage{
deinterlace_file(x)
}
\arguments{
\item{x}{The file}
}
\value{
The corrected character vector
}
\description{
This function removes either \emph{all} odd lines from a file, or \emph{all} even
lines. This can be useful when moving a file between operating systems
(e.g., from Apple to Windows) causes in double line endings.
}
\examples{
metabefor::deinterlace_file(
  "first line",
  "",
  "second (real) line",
  "",
  "third line"
);
}
