% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_named_list.R
\name{df_to_named_list}
\alias{df_to_named_list}
\title{Convert a data frame into a named list}
\usage{
df_to_named_list(df, nameCol = 1, otherCols = 2:(ncol(fd) - 1))
}
\arguments{
\item{df}{The data frame}

\item{nameCol}{The column name or index that holds the names}

\item{otherCols}{The other columns to include (by name or index)}
}
\value{
A named list
}
\description{
Convert a data frame into a named list
}
\examples{
metabefor::df_to_named_list(
  data.frame(
    nameCol = c('name1', 'name2', 'name3'),
    contentCol = c('contents1', 'contents2', 'contents3')
  )
);
}
