% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_data_resources.R
\docType{data}
\name{ebsco_example_1}
\alias{ebsco_example_1}
\title{Simple bibliographic exports from Ebsco}
\format{
A {metabefor} imported search hits object
}
\usage{
ebsco_example_1
}
\description{
A number of simple small datasets with bibliographic information.
}
\details{
These are the datasets obtained by searching in works' titles, at
2024-08-09, for \verb{"evidence synthesis" AND "extraction"} and
for \verb{"evidence synthesis" AND "appraisal"}, in PsycINFO and in MedLine.
}
\keyword{datasets}
