% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_data_resources.R
\docType{data}
\name{example_rxsProject_1}
\alias{example_rxsProject_1}
\title{A minimal Rxs project with two sources extracted in three modules}
\format{
An Rxs project object (class ``)
}
\usage{
example_rxsProject_1
}
\description{
What it says on the tin.
}
\keyword{datasets}
