% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusion_criteria.R
\name{exclusion_criteria}
\alias{exclusion_criteria}
\alias{print.exclusion_criteria}
\title{Specifying exclusion criteria}
\usage{
exclusion_criteria(...)

\method{print}{exclusion_criteria}(x, ...)
}
\arguments{
\item{...}{The exclusion criteria as a series of named
character values.}
}
\value{
A \code{exclusion_criteria} object, basically a dataframe with
two columns: the codes and their descriptions.
}
\description{
Specifying exclusion criteria
}
\examples{
exclusion_criteria(notEmpirical = "Not an empirical study (e.g. a review or debate paper)",
                   notFluffy = "Fluffiness is not assessed");
}
