% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_instructions_from_sheet.R
\name{extractor_instructions_from_sheet}
\alias{extractor_instructions_from_sheet}
\title{Create extractor instructions from a spreadsheet}
\usage{
extractor_instructions_from_sheet(x, headingLevel = 3)
}
\arguments{
\item{x}{The spreadsheet}

\item{headingLevel}{The top-most heading level to use}
}
\value{
The formatted extractor instructions
}
\description{
Create extractor instructions from a spreadsheet
}
