% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenStudies_to_wideDf.R
\name{flattenStudies_to_wideDf}
\alias{flattenStudies_to_wideDf}
\title{Flatten all studies to a wide data frame (somewhat brutally)}
\usage{
flattenStudies_to_wideDf(studies, silent = metabefor::opts$get("silent"))
}
\arguments{
\item{studies}{The studies object (i.e. parsed rxs files)}
}
\value{
A data frame
}
\description{
Flatten all studies to a wide data frame (somewhat brutally)
}
