% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multipleValues.R
\name{get_multipleValues}
\alias{get_multipleValues}
\title{Get values for multiple entities from a Rxs tree, list of trees, or full Rxs project object}
\usage{
get_multipleValues(
  x,
  entityIds = NULL,
  entityIdsRegex = NULL,
  lookInValueLists = TRUE,
  merge_all = TRUE,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{x}{The tree, tree list, or full Rxs project object.}

\item{entityIds}{The entity identifiers of the values to get}

\item{entityIdsRegex}{A regular expression to match again entity identifiers
to obtain the \code{entityIds}.}

\item{lookInValueLists}{Whether to also look inside value lists}

\item{merge_all}{When merging dataframes, if the number of rows is different,
for each source, what to pass as the \code{all} parameter.}

\item{silent}{Whether to be silent or chatty.}

\item{pathString_regex_select}{Regex that the target entities' path strings
have to match (otherwise, the entity is excluded)}
}
\value{
A dataframe
}
\description{
Get values for multiple entities from a Rxs tree, list of trees, or full Rxs project object
}
