% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valueList_asDf.R
\name{get_valueList_asDf_fromStudyTree}
\alias{get_valueList_asDf_fromStudyTree}
\alias{get_valueList_asDf}
\title{Get a value list as data frame from a study tree, list of trees, or studies object}
\usage{
get_valueList_asDf_fromStudyTree(
  x,
  requiredFields = NULL,
  silent = metabefor::opts$get("silent")
)

get_valueList_asDf(
  x,
  requiredFields = NULL,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{x}{The study tree, list of trees, or studies object}

\item{requiredFields}{Fields that have to exist in}

\item{silent}{Whether to be quiet or chatty.}
}
\value{
A dataframe
}
\description{
Get a value list as data frame from a study tree, list of trees, or studies object
}
\examples{

}
