% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSingleValue.R
\name{getSingleValue_fromTree}
\alias{getSingleValue_fromTree}
\alias{getSingleValue_fromTreeList}
\alias{getSingleValue}
\title{Get a single value from a study tree, list of trees, or studies object}
\usage{
getSingleValue_fromTree(
  x,
  entityId,
  lookInValueLists = TRUE,
  silent = metabefor::opts$get("silent")
)

getSingleValue_fromTreeList(x, entityId, returnDf = TRUE)

getSingleValue(x, entityId, returnDf = TRUE)
}
\arguments{
\item{x}{}

\item{entityId}{The entity identifier of the value to get}

\item{lookInValueLists}{Whether to also look inside value lists}

\item{returnDf}{Whether to return a data frame or not}
}
\value{
A list or a dataframe (if \code{returnDf} is \code{TRUE})
}
\description{
Get a single value from a study tree, list of trees, or studies object
}
\examples{

}
