% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IN.R
\name{\%IN\%}
\alias{\%IN\%}
\title{Case insensitive version of \%in\%}
\usage{
find \%IN\% table
}
\arguments{
\item{find}{The element(s) to look up in the vector or matrix.}

\item{table}{The vector or matrix in which to look up the element(s).}
}
\value{
A logical vector.
}
\description{
This is simply 'in', but applies \code{\link[base:toupper]{base::toupper()}} to both
arguments, first.
}
\examples{
letters[1:4] \%IN\% LETTERS

}
