% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_TRUE.R
\name{is_TRUE}
\alias{is_TRUE}
\title{Liberally checking whether a value is TRUE}
\usage{
is_TRUE(x)
}
\arguments{
\item{x}{The value to check.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
This function returns \code{FALSE} if \code{x} is \code{NULL}, \code{NA}, or \code{0}; \code{TRUE} if \code{x}
is a number other than \code{0} or the text \code{TRUE} (case insensitively).
}
