% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_rxsTemplate.R
\name{knit_rxsTemplate}
\alias{knit_rxsTemplate}
\title{Knit the Rxs Template}
\usage{
knit_rxsTemplate(x)
}
\arguments{
\item{x}{The object with the Rxs Specification as produced by a call to
\code{\link[=rxs_fromSpecifications]{rxs_fromSpecifications()}}.}
}
\description{
Knit the Rxs Template
}
