% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsProject_matchingUniqueEntityIdentifiers.R,
%   R/rxsTree_matchingUniqueEntityIdentifiers.R,
%   R/studyTree_matchingUniqueEntityIdentifiers.R
\name{rxsProject_matchingUniqueEntityIdentifiers}
\alias{rxsProject_matchingUniqueEntityIdentifiers}
\alias{rxsTree_matchingUniqueEntityIdentifiers}
\title{Look for all entity identifiers that match a given pattern}
\usage{
rxsProject_matchingUniqueEntityIdentifiers(
  x,
  regex,
  includeValueListsOfMatch = TRUE,
  excludeParentWhenValueListReturned = TRUE
)

rxsTree_matchingUniqueEntityIdentifiers(
  rxsTree,
  regex,
  includeValueListsOfMatch = TRUE,
  excludeParentWhenValueListReturned = TRUE
)

rxsTree_matchingUniqueEntityIdentifiers(
  rxsTree,
  regex,
  includeValueListsOfMatch = TRUE,
  excludeParentWhenValueListReturned = TRUE
)
}
\arguments{
\item{x, rxsTree}{The full Rxs project (i.e. parsed extraction
scripts) or one single Rxs tree.}

\item{regex}{The regular expression to use to find entity identifiers}

\item{includeValueListsOfMatch}{Whether to also include the value lists
inside matching entities (useful for quickly selecting e.g. all
results)}

\item{excludeParentWhenValueListReturned}{Whether, if an entity matches,
has a value list as value, and those value lists are returns (i.e.
\code{includeValueListsOfMatch} is \code{TRUE}), the parent entity (that matched
the regular expression) should be excluded.}
}
\value{
A character vector of identifiers
}
\description{
This function returns all unique entity identifiers of \emph{extractable}
entities that match a given regular expression. This means that identifiers
of container entities or clustering entities are not returned!
}
\examples{
### Load an example Rxs project
data('example_rxsProject_1', package="metabefor");

### Show all entity identifiers
metabefor::rxsProject_get_entityIds(
  example_rxsProject_1
);

### Get all unique identifiers matching "method"
metabefor::rxsProject_matchingUniqueEntityIdentifiers(
  example_rxsProject_1,
  "od"
);
}
