% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTrees.R
\name{mergeTrees}
\alias{mergeTrees}
\title{Merge two \code{data.tree} trees}
\usage{
mergeTrees(tree1, tree2)
}
\arguments{
\item{tree1, tree2}{The trees to merge.}
}
\value{
The merged tree, a (cloned) \code{data.tree} \code{Node} object.
}
\description{
Note this this function is used internally and was written to merge Rxs
trees, so it may not fit other use cases.
}
