% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy_df_to_tree.R
\name{messy_df_to_tree}
\alias{messy_df_to_tree}
\title{Convert a potentially messy 'network data frame' to a data.tree}
\usage{
messy_df_to_tree(x, rootName = "base", childCol = 1, parentCol = 2)
}
\arguments{
\item{x}{The data frame}

\item{rootName}{The name of the root element}

\item{childCol}{The index (as number or name) of the child column}

\item{parentCol}{The index (as number or name) of the parent column}
}
\value{
The data.tree object
}
\description{
Convert a potentially messy 'network data frame' to a data.tree
}
\examples{
messy_df <-
  data.frame(
    childCol = c("A", "B", "C", "D", "E"),
    parentCol = c("B", "B", "A", "", "B")
  );

messy_df_to_tree(
  messy_df
);

### Note that by default, data.tree will deform
### or not accept messy data frames
# data.tree::FromDataFrameNetwork(messy_df);
}
