% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeName.R
\name{nodeName}
\alias{nodeName}
\title{Check whether a prospective node name is valid, sanitize it, and return it}
\usage{
nodeName(
  x,
  entityName = NULL,
  thisEntityText = "this entity (with temporary name '\%s')"
)
}
\arguments{
\item{x}{The prospective node name}

\item{entityName}{The original name of the entity before renaming}
}
\value{
The (sanitized) node name
}
\description{
Check whether a prospective node name is valid, sanitize it, and return it
}
\examples{
metabefor::nodeName("this_is_valid");
metabefor::nodeName("this is sanitized");
if (FALSE) {
  metabefor::nodeName(
    "_thisThrowsAnError:first character can't be _"
  );
  ### This also throws an error: a reserved name by data.tree
  metabefor::nodeName(
    "parent"
  );
}
}
