% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_data_resources.R
\docType{data}
\name{openalex_example_1}
\alias{openalex_example_1}
\alias{openalex_example_2}
\title{Simple bibliographic exports from OpenAlex}
\format{
A data frame
}
\usage{
openalex_example_1
}
\description{
A number of simple small datasets with bibliographic information.
}
\details{
These are the datasets obtained by searching,
in works' titles, for respectively:
\itemize{
\item \code{openalex_example_1}: 2024-08-09, \verb{"evidence synthesis" AND "extraction"},
\url{https://openalex.org/works?filter=display_name.search\%3A\%22evidence\%20synthesis\%22\%20AND\%20\%22extraction\%22,type\%3A\%21types\%2Fdataset\%7Ctypes\%2Flibguides\%7Ctypes\%2Fsupplementary-materials}
\item \code{openalex_example_2}: 2024-08-09, \verb{"evidence synthesis" AND "appraisal"},
\url{https://openalex.org/works?filter=display_name.search\%3A\%22evidence\%20synthesis\%22\%20AND\%20\%22appraisal\%22,type\%3A\%21types\%2Fdataset\%7Ctypes\%2Flibguides\%7Ctypes\%2Fsupplementary-materials}
}
}
\keyword{datasets}
