% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override.R
\name{override}
\alias{override}
\title{Override default values with bespoke alternatives}
\usage{
override(default, bespoke)
}
\arguments{
\item{default}{The default values}

\item{bespoke}{The bespoke alternatives}
}
\value{
A vector that is the union of \code{default} and \code{bespoke}, but for
all elements that occur in both, the version from \code{bespoke}.
}
\description{
This function takes two names lists or two named vectors and replaces
all value with a bespoke specification with that bespoke specification.
}
\examples{
override(
  default = c(a = "banana",
              b = "grape",
              c = "pineapple"),
  bespoke = c(b = "peach",
              c = "apple")
);
}
