% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padVectors.R
\name{padVectors}
\alias{padVectors}
\title{In a list of vectors, pad short vectors}
\usage{
padVectors(x, padWith = NA)
}
\arguments{
\item{x}{The list of vectors}

\item{padWith}{The element to pad with}
}
\value{
The list with the padded vectors
}
\description{
In a list of vectors, pad short vectors
}
\examples{
metabefor::padVectors(
  list(
    c(1, 2, 3),
    4,
    c(5, 6)
  )
);
}
