% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rxs.R
\name{print.rxs}
\alias{print.rxs}
\title{Title}
\usage{
\method{print}{rxs}(studyTree, rxsStructure = NULL, knit = TRUE, headerPrefix = "### ", ...)
}
\arguments{
\item{studyTree}{}

\item{rxsStructure}{}

\item{knit}{}

\item{headerPrefix}{}

\item{...}{}
}
\value{

}
\description{
Title
}
