% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rxs.R
\name{rxs_partial}
\alias{rxs_partial}
\alias{knit_print.rxs}
\alias{print.rxs}
\title{Printing a tree of R extraction scripts}
\usage{
rxs_partial(
  studyTree,
  rxsStructure = NULL,
  headingLevel = x$headingLevel,
  printPlots = TRUE,
  echoPartial = FALSE,
  partialFile = NULL,
  ...
)

\method{knit_print}{rxs}(
  studyTree,
  rxsStructure = NULL,
  headingLevel = 3,
  echoPartial = FALSE,
  partialFile = NULL,
  ...
)

\method{print}{rxs}(
  studyTree,
  rxsStructure = NULL,
  headingLevel = 3,
  forceKnitrOutput = FALSE,
  echoPartial = FALSE,
  partialFile = NULL,
  ...
)
}
\arguments{
\item{studyTree}{The study tree}

\item{rxsStructure}{Optionally, the rxs structure}

\item{...}{}

\item{headerPrefix}{}
}
\value{
invisibly, the dataframe with paths and values.
}
\description{
Printing a tree of R extraction scripts
}
