% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aggregationTree_from_gs.R
\name{read_aggregationTree_from_gs}
\alias{read_aggregationTree_from_gs}
\title{Read an aggregation tree from a Google Sheets URL}
\usage{
read_aggregationTree_from_gs(aggregationTreeURL)
}
\arguments{
\item{aggregationTreeURL}{The URL (make sure it's publicly viewable)}
}
\value{
The data.tree object
}
\description{
Read an aggregation tree from a Google Sheets URL
}
