% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderTreeAsGraph.R
\name{renderTreeAsGraph}
\alias{renderTreeAsGraph}
\title{Render a tree as a graph}
\usage{
renderTreeAsGraph(tree, theme = metabefor::defaultGraphTheme())
}
\arguments{
\item{tree}{The tree}

\item{theme}{The theme}
}
\value{
The result of \code{DiagrammeR::render_graph()}.
}
\description{
Render a tree as a graph
}
