% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_buildTemplate.R
\name{rxs_buildTemplate}
\alias{rxs_buildTemplate}
\title{Title}
\usage{
rxs_buildTemplate(
  rxsStructure,
  yamlMetadata = list(title = "Systematic Review Extraction Script Template", author =
    NULL, date = format(Sys.time(), "\%d \%b \%Y at \%H:\%M:\%S")),
  gs_url = NULL,
  indent = TRUE,
  indentSpaces = 2,
  fullWidth = 80,
  commentCharacter = "#",
  fillerCharacter = "#",
  eC = entityColNames(),
  repeatingSuffix = "__1__",
  silent = FALSE
)
}
\arguments{
\item{rxsStructure}{}

\item{yamlMetadata}{}

\item{gs_url}{}

\item{indent}{}

\item{indentSpaces}{}

\item{fullWidth}{}

\item{commentCharacter}{}

\item{fillerCharacter}{}

\item{eC}{}

\item{repeatingSuffix}{}

\item{silent}{}
}
\value{

}
\description{
Title
}
