% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_fromSpecifications.R
\name{rxs_fromSpecifications}
\alias{rxs_fromSpecifications}
\title{Title}
\usage{
rxs_fromSpecifications(
  gs_url = NULL,
  ws = list(entities = "entities", valueTemplates = "valueTemplates", definitions =
    "definitions", instructions = "instructions"),
  entitiesFilename = NULL,
  valueTemplatesFilename = NULL,
  definitionsFilename = NULL,
  localBackup = list(entities = NULL, valueTemplates = NULL, definitions = NULL),
  outputFile = NULL,
  yamlMetadata = list(title = "Systematic Review Extraction Script Template", author =
    NULL, date = format(Sys.time(), "\%d \%b \%Y at \%H:\%M:\%S")),
  author = NULL,
  indent = TRUE,
  indentSpaces = 2,
  fullWidth = 80,
  commentCharacter = "#",
  fillerCharacter = "#",
  eC = metabefor::opts$get(entityColNames),
  valueTemplateCols = metabefor::opts$get(valueTemplateColNames),
  repeatingSuffix = "__1__",
  rootName = "study",
  silent = FALSE,
  instructionHeadingLevel = 3,
  graphTheme = list(c("fontname", "Arial", "node")),
  returnFullObject = TRUE
)
}
\arguments{
\item{gs_url}{}

\item{ws}{}

\item{entitiesFilename}{}

\item{valueTemplatesFilename}{}

\item{definitionsFilename}{}

\item{localBackup}{}

\item{outputFile}{}

\item{yamlMetadata}{}

\item{author}{}

\item{indent}{}

\item{indentSpaces}{}

\item{fullWidth}{}

\item{commentCharacter}{}

\item{fillerCharacter}{}

\item{eC}{}

\item{valueTemplateCols}{}

\item{repeatingSuffix}{}

\item{rootName}{}

\item{silent}{}

\item{returnFullObject}{}
}
\value{

}
\description{
Title
}
