% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_data_resources.R
\docType{data}
\name{rxs_minimal_example}
\alias{rxs_minimal_example}
\alias{rxs_minimal_modular_example}
\title{An very rudimentary example Rxs specification}
\format{
An example of an Rxs specification
}
\usage{
rxs_minimal_example
}
\description{
This is a simple and relatively short Rxs specification. One version is
parsed into a single Rxs template, and the other into three modules.
}
\keyword{datasets}
