% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_parseExtractionScripts.R
\name{rxs_parseExtractionScripts}
\alias{rxs_parseExtractionScripts}
\title{Read and parse the extraction scripts in a directory}
\usage{
rxs_parseExtractionScripts(
  path,
  pattern = "\\\\.rxs\\\\.Rmd$|\\\\.rxs$",
  exclude = c("extractionScriptTemplate.rxs.Rmd", "\\\\[EXCLUDED]"),
  ignore.case = TRUE,
  recursive = TRUE,
  silent = metabefor::opts$get("silent"),
  showErrors = TRUE,
  encoding = "UTF-8"
)
}
\arguments{
\item{path}{The path containing the extraction scripts.}

\item{pattern}{The regular expression specifying which files to read.}

\item{exclude}{A regular expression specifying which files to ignore.}

\item{ignore.case}{Whether the regular expression is case sensitive.}

\item{recursive}{Whether to also look in subdirectories.}

\item{silent}{Whether to be quiet or chatty.}

\item{showErrors}{Whether to show or hide errors that are encountered.}

\item{encoding}{The files' encoding.}
}
\value{
A list of parsed extraction scripts.
}
\description{
Read and parse the extraction scripts in a directory
}
