% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_parseSingleExtractionScript.R
\name{rxs_parseSingleExtractionScript}
\alias{rxs_parseSingleExtractionScript}
\title{Read and parse one single extraction script}
\usage{
rxs_parseSingleExtractionScript(
  filename,
  path,
  progress = NULL,
  silent = metabefor::opts$get("silent"),
  parallel_options = NULL,
  showErrors = TRUE,
  encoding = "UTF-8"
)
}
\arguments{
\item{filename}{The filename with the extraction script.}

\item{path}{The path to the directory with that extraction script.}

\item{progress}{Optionally, a progress bar to tick after the file
has been parsed  (must be a \code{\link[progress:progress_bar]{progress::progress_bar()}} object.}

\item{silent}{Whether to be quiet or chatty.}

\item{parallel_options}{Used to pass options for parallel processing.}

\item{showErrors}{Whether to show or hide errors that are encountered.}

\item{encoding}{The files' encoding.}
}
\value{
A list of parsed extraction scripts.
}
\description{
Read and parse one single extraction script
}
