% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_parseSpecifications.R
\name{rxs_parseSpecifications}
\alias{rxs_parseSpecifications}
\title{Title}
\usage{
rxs_parseSpecifications(
  entities,
  valueTemplates,
  definitions = NULL,
  eC = metabefor::opts$get(entityColNames),
  valueTemplateCols = metabefor::opts$get(valueTemplateColNames),
  rootName = "study"
)
}
\arguments{
\item{entities}{}

\item{valueTemplates}{}

\item{definitions}{}

\item{eC}{}

\item{valueTemplateCols}{}

\item{rootName}{}
}
\value{

}
\description{
Title
}
