% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_buildModularTemplate.R,
%   R/rxs_buildTemplate.R, R/rxs_fromSpecifications.R
\name{rxs_buildModularTemplate}
\alias{rxs_buildModularTemplate}
\alias{rxs_buildTemplate}
\alias{rxs_fromSpecifications}
\title{Generate an Rxs template}
\usage{
rxs_buildModularTemplate(
  rxsStructure,
  module,
  yamlMetadata = list(title = "Systematic Review Extraction Script Template", author =
    NULL, date = format(Sys.time(), format(Sys.time(),
    "\%Y-\%m-\%d at \%H:\%M:\%S \%Z (UTC\%z)"))),
  gs_url = NULL,
  indent = TRUE,
  indentSpaces = 2,
  fullWidth = 80,
  commentCharacter = "#",
  fillerCharacter = "#",
  eC = metabefor::opts$get("entityColNames"),
  repeatingSuffix = "__1__",
  silent = FALSE
)

rxs_buildTemplate(
  rxsStructure,
  yamlMetadata = list(title = "Systematic Review Extraction Script Template", author =
    NULL, date = format(Sys.time(), "\%d \%b \%Y at \%H:\%M:\%S")),
  gs_url = NULL,
  indent = TRUE,
  indentSpaces = 2,
  fullWidth = 80,
  commentCharacter = "#",
  fillerCharacter = "#",
  eC = metabefor::opts$get("entityColNames"),
  repeatingSuffix = "__1__",
  silent = FALSE
)

rxs_fromSpecifications(
  x = NULL,
  outputFile = NULL,
  outputPath = NULL,
  outputFilenamePattern = NULL,
  localBackup = NULL,
  yamlMetadata = list(title = "Systematic Review Extraction Script Template", author =
    NULL, date = format(Sys.time(), "\%Y-\%m-\%d at \%H:\%M:\%S \%Z (UTC\%z)")),
  indent = TRUE,
  indentSpaces = 2,
  fullWidth = 80,
  commentCharacter = "#",
  fillerCharacter = "#",
  ws = metabefor::opts$get("rxsSheetnames"),
  eC = metabefor::opts$get("entityColNames"),
  valueTemplateCols = metabefor::opts$get("valueTemplateColNames"),
  instructionsCols = metabefor::opts$get("instructionsColNames"),
  definitionsCols = metabefor::opts$get("definitionsColNames"),
  repeatingSuffix = "__1__",
  rootName = "study",
  silent = metabefor::opts$get("silent"),
  instructionHeadingLevel = 3,
  graphTheme = list(c("fontname", "Arial", "node")),
  ignoreModules = FALSE,
  returnFullObject = TRUE,
  gs_url = NULL,
  localFile = NULL
)
}
\arguments{
\item{rxsStructure}{The Rxs specification as loaded from a spreadsheet.}

\item{yamlMetadata}{A names list containing character values named
\code{title}, \code{author}, and \code{date} which will be set in the Rmd file that is
the Rxs template as metadata.}

\item{gs_url}{DEPRECATED - please use \code{x} instead!
A Google Sheets
URL (make sure it's viewable by anybody with the link!).}

\item{indent}{Whether to use indentation to visually organise the Rxs
template. If TRUE, deeper nesting in the Rxs specification's hierarchy
will be visible as deeper indentation.}

\item{indentSpaces}{The number of spaces to use when identing.}

\item{fullWidth}{The maximum width of the Rxs template in characters.}

\item{commentCharacter}{The character used to signify comments - if this is
changed, R will throw errors (unless perhaps it once introduces another
comment symbol).}

\item{fillerCharacter}{The character used after the first character for
filling up space.}

\item{eC}{The entity columns; a named list with character values holding the
names of the columns in the \code{entities} worksheet of the spreadsheet. The
default values are stored in \code{metabefor::opts$get("entityColNames")} - if you
need to override these values, just reproduce that object.}

\item{repeatingSuffix}{The suffix to use for the entity identifiers/names
of repeating entities.}

\item{silent}{Whether to be silent or chatty.}

\item{x}{A Google Sheets URL (make sure it's viewable by anybody
with the link!) or either a character
value with a valid path to a spreadsheet (\code{.xlsx}) with the worksheets, or
a named list of character strings each pointing to a \code{.csv} file with the
relevant worksheet, where the names of each character value represent the
names of the worksheets contained in each file.}

\item{outputFile}{When not working with modules, the path to a directory
and file where the Rxs template will be stored. If \code{NULL}, the produced
Rxs template (or the full object, depending on \code{returnFullObject}) is
returned visibly.}

\item{outputPath, outputFilenamePattern}{If \code{NULL}, the produced Rxs
templates are returned (visibly) in a list (or the full object is, depending
on \code{returnFullObject}). If not \code{NULL}, \code{outputPath} has
to be a valid path where the Rxs templates will be written to, and
\code{outputFilenamePattern} has to be a valid character value (i.e. a text
string) containing \verb{\%s} in the position where the module identifier will
be inserted in the filenames.}

\item{localBackup}{If not \code{NULL}, either a path to a single \code{.xlsx} file to
save a local backup spreadsheet to, or a named list with character values,
where each name is the name of a worksheet (e.g. \code{entities},
\code{valueTemplates}, \code{definitions}, and \code{instructions}), and each named value
is the path to store the corresponding worksheets will be stored as \code{.csv}
file.}

\item{ws}{The worksheet names: a named list with four character values
named \code{entities}, \code{valueTemplates}, \code{definitions}, and \code{instructions}.}

\item{valueTemplateCols}{The value template columns; a named list with
character values holding the names of the columns in the \code{entities}
worksheet of the spreadsheet. The default values are stored
in \code{metabefor::opts$get("valueTemplateColNames")} - if you need to
override these values, just reproduce that object.}

\item{rootName}{The name of the root element}

\item{instructionHeadingLevel}{The top-most heading level for the
instructions.}

\item{graphTheme}{The graph theme to use.}

\item{ignoreModules}{Optionally, you can ignore modules specified in the
Rxs specification spreadsheet by setting \code{ignoreModules} to \code{TRUE}.}

\item{returnFullObject}{Whether to return the full object or just the
template.}

\item{localFile}{DEPRECATED - please use \code{x} instead!
When reading from a local file, pass either a character
value with a valid path to a spreadsheet (\code{.xlsx}) with the worksheets, or
a named list of character strings each pointing to a \code{.csv} file with the
relevant worksheet, where the names of each character value represent the
names of the worksheets contained in each file.}
}
\value{
Either a full object, containing the template and other products,
or just the template.
}
\description{
These functions process an R extraction script template specification in
a spreadsheet format and produce the corresponding R extraction script
template file (an R Markdown file with the extension \code{.rxs.Rmd}).
\code{rxs_fromSpecifications()} can be passed a Google Sheet URL, and uses
\code{rxs_buildTemplate()} in the background to produce the template.
}
