% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_validation.R
\name{rxs_validation}
\alias{rxs_validation}
\title{Validate a study tree}
\usage{
rxs_validation(studyTree, rxsStructure = NULL)
}
\arguments{
\item{studyTree}{The study tree}

\item{rxsStructure}{Optionally, the \code{rxsStructure} as resulting from
a call to \code{\link[=rxs_parseSpecifications]{rxs_parseSpecifications()}}.}
}
\value{
Invisibly, the study tree (which was altered in place, consistent
with the reference semantics employed by \code{\link[data.tree:Node]{data.tree::Node()}}.
}
\description{
Validate a study tree
}
