% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsProject_to_valueDf.R
\name{rxsProject_to_valueDf}
\alias{rxsProject_to_valueDf}
\title{Create a long ('tidy') data frame with the full Rxs tree from each source}
\usage{
rxsProject_to_valueDf(rxsObject, zapNewlines = TRUE)
}
\arguments{
\item{zapNewlines}{Whether to call \code{\link[=zapNewlines]{zapNewlines()}} on the values}

\item{rxsTree}{The Rxs tree}
}
\value{
A data frame
}
\description{
The data frame has columns "\code{sourceId}", "\code{path}", "\code{entity}", and "\code{nodeValue}".
}
