% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsTree_to_valueDf.R
\name{rxsTree_to_valueDf}
\alias{rxsTree_to_valueDf}
\title{Create a long ('tidy') data frame with the full Rxs tree}
\usage{
rxsTree_to_valueDf(rxsTree, zapNewlines = TRUE)
}
\arguments{
\item{rxsTree}{The Rxs tree}

\item{zapNewlines}{Whether to call \code{\link[=zapNewlines]{zapNewlines()}} on the values}
}
\value{
A data frame
}
\description{
The data frame has columns "\code{path}", "\code{entity}", and "\code{nodeValue}".
}
