% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_for_DiagrammeR.R
\name{sanitize_for_DiagrammeR}
\alias{sanitize_for_DiagrammeR}
\title{Sanitize for DiagrammeR}
\usage{
sanitize_for_DiagrammeR(
  x,
  regExReplacements = metabefor::opts$get("diagrammerSanitization")
)
}
\arguments{
\item{x}{The string or vector}

\item{regExReplacements}{A list of two-element character vectors; first element
should be the element to search, and the second element, the replacement.}
}
\value{
The sanitized character vector
}
\description{
Basically a wrapper for \code{\link[=gsub]{gsub()}} to sanitize a string for
DiagrammeR
}
\examples{
metabefor::sanitize_for_DiagrammeR("This is or isn't problematic");
}
