% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_or_load.R
\name{save_or_load_rds}
\alias{save_or_load_rds}
\title{Something akin to caching}
\usage{
save_or_load_rds(
  object,
  path,
  prefix = "autogenerated",
  suffix = "",
  filesToKeep = 2
)
}
\arguments{
\item{object}{The object to save or load.}

\item{path}{The directory where to save or load the object (from).}

\item{prefix, suffix}{A prefix or suffix to add in the filename.}

\item{filesToKeep}{The number of files to keep.}
}
\value{
The object.
}
\description{
This function check whether an object exists; if it does, it saves
it using a system where the number of backups is limited (to 2 files
by default), so automatically deleting the oldest version; if it does
not exist, it loads one of the automatically saved versions. This
functionality can be useful when working with large scripts that
take a lot of time to create or process objects.
}
