% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rxsTree_in_rxsStructure.R
\name{show_rxsTree_in_rxsStructure}
\alias{show_rxsTree_in_rxsStructure}
\title{Show an imported Rxs structure}
\usage{
show_rxsTree_in_rxsStructure(
  x,
  output = NULL,
  outputModuleFilename = "extraction-tree--simple--\%s.pdf",
  headingLevel = 3
)
}
\arguments{
\item{x}{The parsed Rxs specifications as read
with \code{\link[=rxs_fromSpecifications]{rxs_fromSpecifications()}}.}

\item{output, outputModuleFilename}{When not NULL, pass a path to a
filename where the RxsTree should be written to. If multiple modules were
read, the filename is determine by \code{outputModuleFilename}, where \verb{\%s} will
be replaced with the module name.}

\item{headingLevel}{The level of the heading to use is multiple modules
were specified.}
}
\value{
Invisibly; or, when knitting, an object that knitr will print.
}
\description{
This function shows the Rxs structure that was imported from
an Rxs specification
}
