% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_search_hits_by_database.R
\name{show_search_hits_by_database}
\alias{show_search_hits_by_database}
\title{Show the search hits by database}
\usage{
show_search_hits_by_database(x, output = "kable")
}
\arguments{
\item{x}{The object with search results, as produced by a call to
\code{\link[=import_search_results]{import_search_results()}}.}

\item{output}{Output format: \code{"kable"} or \code{"table"}.}
}
\description{
Show the search hits by database
}
