% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_search_hits_with_dois.R
\name{show_search_hits_with_dois}
\alias{show_search_hits_with_dois}
\title{Show the number of search hits with DOIs}
\usage{
show_search_hits_with_dois(x, output = "kable")
}
\arguments{
\item{x}{The object with search results, as produced by a call to
\code{\link[=import_search_results]{import_search_results()}}.}

\item{output}{Output format: \code{"kable"} or \code{"table"}.}
}
\description{
Show the number of search hits with DOIs
}
