% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_save_or_load_rds.R,
%   R/save_or_load_csv.R, R/save_or_load_rds.R
\name{create_and_save_or_load_rds}
\alias{create_and_save_or_load_rds}
\alias{save_or_load_csv}
\alias{save_or_load_rds}
\title{Something akin to caching}
\usage{
create_and_save_or_load_rds(
  tryToSave,
  expression,
  objectName,
  path,
  prefix = "autogenerated",
  suffix = "",
  filesToKeep = 2,
  silent = metabefor::opts$get("silent")
)

save_or_load_csv(
  object,
  path,
  tryToSave = TRUE,
  prefix = "autogenerated",
  suffix = "",
  filesToKeep = 2,
  silent = metabefor::opts$get("silent")
)

save_or_load_rds(
  object,
  path,
  tryToSave = TRUE,
  prefix = "autogenerated",
  suffix = "",
  filesToKeep = 2,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{tryToSave}{The condition to check.}

\item{expression}{The expression to execute.}

\item{objectName}{The object to save the expression's result to.}

\item{path}{The directory where to save or load the object (from).}

\item{prefix, suffix}{A prefix or suffix to add in the filename.}

\item{filesToKeep}{The number of files to keep.}

\item{silent}{Whether to be chatty or silent.}

\item{object}{The object to save or load.}
}
\value{
The object.
}
\description{
These functions try to facilitate working with large objects.
\code{save_or_load_rds} and \code{save_or_load_csv} check whether an object exists;
if it does and \code{tryToSave} is TRUE, it saves
it using a system where the number of backups is limited (to 2 files
by default), so automatically deleting the oldest version; and f it does
not exist, it loads one of the automatically saved versions. This
functionality can be useful when working with large scripts that
take a lot of time to create or process objects.
\code{create_and_save_or_load_rds} checks \code{tryToSave}; if it's \code{TRUE}, it
executes the expression and saves the result; otherwise, it tries to load
the file from disk.
}
