% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_to_freqTab.R,
%   R/studyTreeList_to_freqTab.R, R/studyTree_to_freqTab.R
\name{studies_to_freqTab}
\alias{studies_to_freqTab}
\alias{studyTreeList_to_freqTab}
\alias{studyTree_to_freqTab}
\title{Create a frequency table from a list of study trees}
\usage{
studies_to_freqTab(
  x,
  rowRegex,
  colRegex,
  rowTargetValue = NULL,
  colTargetValue = NULL,
  fillValue = 0,
  rowColMultiplicationFunction = `*`,
  rowTargetFunction = `==`,
  colTargetFunction = `==`,
  aggregationFunction = `+`,
  rowLabels = NULL,
  colLabels = NULL,
  flattenValues = flattenValues,
  includeValueListsOfMatch = TRUE,
  excludeParentWhenValueListReturned = TRUE,
  silent = metabefor::opts$get("silent")
)

studyTreeList_to_freqTab(
  x,
  rowRegex,
  colRegex,
  rowTargetValue = NULL,
  colTargetValue = NULL,
  fillValue = 0,
  rowColMultiplicationFunction = `*`,
  rowTargetFunction = `==`,
  colTargetFunction = `==`,
  aggregationFunction = `+`,
  rowLabels = NULL,
  colLabels = NULL,
  flattenValues = TRUE,
  includeValueListsOfMatch = TRUE,
  excludeParentWhenValueListReturned = TRUE,
  silent = metabefor::opts$get("silent")
)

studyTree_to_freqTab(
  x,
  rowRegex,
  colRegex,
  rowTargetValue = NULL,
  colTargetValue = NULL,
  valuePreprocessingFunction = vectorValue_to_valueList,
  rowColMultiplicationFunction = `*`,
  rowTargetFunction = `==`,
  colTargetFunction = `==`,
  rowLabels = NULL,
  colLabels = NULL,
  rowOrder = NULL,
  colOrder = NULL,
  sortRowsAlphabetically = TRUE,
  sortColsAlphabetically = TRUE,
  includeValueListsOfMatch = TRUE,
  flattenValues = TRUE,
  excludeParentWhenValueListReturned = TRUE,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{x}{}

\item{rowRegex, colRegex}{Regular expressions used to find the entities
that will form the rows or columns}

\item{rowTargetValue, colTargetValue}{Value to consider a 'hit' for the
row and column entities}

\item{fillValue}{The value to insert in rows or columns that have to
be added because of inconsistencies between extraction scripts}

\item{rowColMultiplicationFunction}{The function to use to multiply the
rows and columns into the matrix}

\item{rowTargetFunction, colTargetFunction}{Functions to use to compare,
the entity values found for the rows and columns to the target
values (e.g. \code{==}, \code{>}, \code{<}, etc).}

\item{aggregationFunction}{The function to use to aggregate matrices}

\item{rowLabels, colLabels}{A names vector used to replace row and column
labels; the indices (element names) should be the entity identifiers and
the values the labels to use.}

\item{flattenValues}{Whether to flatten values that are vectors into
single string values that represents those values, or not (in which
case they remain vectors, and so single studies occur multiple times
in the frequency table).}

\item{includeValueListsOfMatch}{Whether to also include the value lists
inside matching entities (useful for quickly selecting e.g. all
results)}

\item{excludeParentWhenValueListReturned}{Whether, if an entity matches,
has a value list as value, and those value lists are returns (i.e.
\code{includeValueListsOfMatch} is \code{TRUE}), the parent entity (that matched
the regular expression) should be excluded.}
}
\value{

}
\description{
Create a frequency table from a list of study trees
}
