% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyTree_has_entity.R
\name{studyTree_has_entity}
\alias{studyTree_has_entity}
\title{Check whether a study tree has an entity}
\usage{
studyTree_has_entity(x, entityId_regex)
}
\arguments{
\item{x}{The study tree.}

\item{entityId_regex}{The regular expression}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Check whether a study tree has an entity with an identifier that
matches the regular expression.
}
