% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyTree_to_dataframe.R
\name{studyTree_to_valueDf}
\alias{studyTree_to_valueDf}
\title{Create a long ('tidy') data frame with the full study tree}
\usage{
studyTree_to_valueDf(studyTree)
}
\arguments{
\item{studyTree}{The study tree}
}
\value{
A data frame
}
\description{
The data frame has columns "\code{path}", "\code{entity}", and "\code{nodeValue}".
}
