% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_data_from_list.R
\name{supplement_data_from_list}
\alias{supplement_data_from_list}
\title{Title}
\usage{
supplement_data_from_list(
  studyTree,
  targetNode_id,
  targetNode_sourceNodeIdField,
  idField_in_sourceNode = NULL,
  idField_in_targetNode = NULL
)
}
\arguments{
\item{studyTree}{THe study tree}

\item{targetNode_id}{The identifier of the target node (the node to
supplement)}

\item{targetNode_sourceNodeIdField}{Inside the target node, the field name
holding that identifier of the source node (the node supplying the data)}

\item{idField_in_sourceNode}{\code{NULL} if the source node's name is also its
identifier; otherwise, the name of the field containing the identifier in
the source node.}

\item{idField_in_TargetNode}{\code{NULL} if the target node's name is also its
identifier; otherwise, the name of the field containing the identifier in
the target node.}
}
\value{

}
\description{
Title
}
