% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_studyTree.R, R/tidy_studyTrees.R
\name{tidy_studyTree}
\alias{tidy_studyTree}
\alias{tidy_studyTrees}
\title{Flatten one or more study trees into a tidy data frame}
\usage{
tidy_studyTree(
  studyTree,
  studyName = "",
  eC = metabefor::opts$get("entityColNames")
)

tidy_studyTrees(listOfStudyTrees, eC = metabefor::opts$get("entityColNames"))
}
\arguments{
\item{studyTree}{The study tree to convert}

\item{studyName}{The name of the study}

\item{eC}{The names of the entity columns}

\item{listOfStudyTrees}{A list of study trees}
}
\value{
A data frame
}
\description{
Flatten one or more study trees into a tidy data frame
}
