% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_entityValue.R
\name{transform_entityValue}
\alias{transform_entityValue}
\title{Perform a transformation selectively}
\usage{
transform_entityValue(
  x,
  fun,
  entityValue_argName = "x",
  entityId_regex = NULL,
  newEntityName_prefix = "",
  newEntityName_suffix = "_trfmd",
  funArgs = NULL,
  requiredField_regex = NULL
)
}
\arguments{
\item{x}{The full Rxs project object (as produced
by \code{\link[=rxs_parseExtractionScripts]{rxs_parseExtractionScripts()}}).}

\item{fun}{The function to apply.}

\item{entityValue_argName}{The argument name to pass the entity value as.}

\item{entityId_regex}{An optional regular expression: if specified, only
entity nodes with entity identifiers that match this regular expression will
be processed.}

\item{newEntityName_prefix, newEntityName_suffix}{The prefix and suffix to
add to the entity identifier to create the name of the new entity that is
created. That entity is added as a sibling of the target entity.}

\item{funArgs}{Additional arguments, as a names list with each element's
name is the argument name and the element itself the content.}

\item{requiredField_regex}{This functionality has not yet been implemented.}
}
\value{
Invisibly, the full Rxs project object. Note that the Rxs trees
will be changed in place given \code{data.tree}'s pass-by-reference logic; so
you can discard the result.
}
\description{
\code{\link[=transform_entityValue]{transform_entityValue()}} takes a full Rxs project object (as
produced by \code{\link[=rxs_parseExtractionScripts]{rxs_parseExtractionScripts()}}) and processes all
Rxs trees, looking for entities that match the \code{entityId_regex}
regular expression (for information about what clustering entities are,
see \url{https://sysrevving.com/glossary.html})
and passes those to function \code{fun}, passing the entity's value as the
argument named in \code{entityValue_argName} and with \code{funArgs} as additional
arguments.
}
\examples{
### Load an example Rxs project
data('example_rxsProject_1', package="metabefor");

### Show the sample sizes
metabefor::get_singleValue(
  example_rxsProject_1,
  "sampleSize"
);

### Double the sample sizes
metabefor::transform_entityValue(
  example_rxsProject_1,
  function(x) {
    return(x * 2);
  },
  entityId_regex = "sampleSize"
);

### Show transformed result

### Show the sample sizes
metabefor::get_singleValue(
  example_rxsProject_1,
  "sampleSize_trfmd"
);
}
