% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsProject_vectorValue_to_valueList.R,
%   R/rxsTreeList_vectorValue_to_valueList.R,
%   R/rxsTree_vectorValue_to_valueList.R, R/vectorValue_to_valueList.R
\name{rxsProject_vectorValue_to_valueList}
\alias{rxsProject_vectorValue_to_valueList}
\alias{rxsTreeList_vectorValue_to_valueList}
\alias{rxsTree_vectorValue_to_valueList}
\alias{vectorValue_to_valueList}
\title{In an entity, convert a vector to a value list}
\usage{
rxsProject_vectorValue_to_valueList(x, entityId, allPossibleValues = NULL)

rxsTreeList_vectorValue_to_valueList(x, entityId, allPossibleValues = NULL)

rxsTree_vectorValue_to_valueList(x, entityId, allPossibleValues = NULL)

vectorValue_to_valueList(
  value,
  allPossibleValues = NULL,
  labels = NULL,
  valueForPresence = 1,
  valueForAbsence = 0
)
}
\arguments{
\item{x}{A Rxs tree}

\item{entityId}{The entity Id}

\item{allPossibleValues}{A list of all possible values}

\item{value}{The vector value to convert}

\item{labels}{Optionally, a named vector with labels to replace the actual
values with; names are values, values are labels.}

\item{valueForPresence}{Value to assign for the elements in the vector}

\item{valueForAbsence}{Value to assign for the elements
in \code{allPossibleValues} that do not occur in the vector}
}
\value{
The Rxs tree specified in \code{x}, invisibly.

The list
}
\description{
For example, \code{c('red', 'brown')} becomes \code{list(red = 1, brown = 1)}.
}
\examples{
### Load an example Rxs project
data('example_rxsProject_1', package="metabefor");

### Look at the sample size specified in
### an Rxs tree for a source
example_rxsProject_1$rxsTrees$qurid_7h50rzmz$methods$sample$sampleSize$value;

### Set it to a vector
example_rxsProject_1$rxsTrees$qurid_7h50rzmz$methods$sample$sampleSize$value <-
  c(80, 62);

### Convert the values to a value list
rxsTree_vectorValue_to_valueList(
  example_rxsProject_1$rxsTrees$qurid_7h50rzmz,
  "sampleSize"
);

### Verify that this happened
example_rxsProject_1$rxsTrees$qurid_7h50rzmz$methods$sample$sampleSize$value;

}
