% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_extractor_instructions.R
\name{write_extractor_instructions}
\alias{write_extractor_instructions}
\title{Write the extractor instructions}
\usage{
write_extractor_instructions(
  x,
  output = NULL,
  headingLevel = metabefor::opts$get("defaultHeadingLevel")
)
}
\arguments{
\item{x}{An rxsStructure object, as imported
by metabefor::rxs_fromSpecifications().}

\item{output}{The output file to write to.}
}
\description{
Write the extractor instructions
}
