% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_JabRef_Config.r
\name{write_JabRef_Config}
\alias{write_JabRef_Config}
\title{Write a JabRef configuration file}
\usage{
write_JabRef_Config(
  outputPath,
  screeners = c("a", "b"),
  screenerFieldsPrefix = "screener",
  screenerFieldsSuffix = "status",
  screenerConfidencePrefix = "screener",
  screenerConfidenceSuffix = "confidence",
  fields = c("title", "abstract"),
  duplicateField = NULL,
  sortField = "title",
  sortDesc = FALSE,
  screeningType = "screening",
  jabrefVersion = "2.11.1",
  jabrefXML = NULL,
  batFile = NULL,
  generateSettingRemovalBatchFile = FALSE,
  screenerSuffixInFilename = FALSE
)
}
\arguments{
\item{outputPath}{}

\item{screeners}{}

\item{screenerFieldsPrefix}{}

\item{screenerFieldsSuffix}{}

\item{screenerConfidencePrefix}{}

\item{screenerConfidenceSuffix}{}

\item{fields}{}

\item{duplicateField}{}

\item{sortField}{}

\item{sortDesc}{}

\item{screeningType}{}

\item{jabrefVersion}{}

\item{jabrefXML}{}

\item{batFile}{}

\item{generateSettingRemovalBatchFile}{}

\item{screenerSuffixInFilename}{}
}
\value{

}
\description{
Write a JabRef configuration file
}
