% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_screenerPackage.r
\name{write_screenerPackage}
\alias{write_screenerPackage}
\title{Write a 'screener package' to disk}
\usage{
write_screenerPackage(
  bibliographyDf,
  outputPath,
  screeners = c("a", "b"),
  screenerFieldsPrefix = "screener",
  screenerFieldsSuffix = "status",
  prevRoundField = NULL,
  duplicateField = NULL,
  initializeScreenerFields = TRUE,
  basename = "screening_",
  silent = metabefor::opts$get("silent"),
  ...
)
}
\arguments{
\item{bibliographyDf}{A dataframe with bibliographic entries}

\item{outputPath}{The path where to write the screener package(s)}

\item{screeners}{The unique identifiers to use for each screener; a character
vector where each value represented one screener. This will be used for the
directory and filenames as well as fields in the bibliographic database.}

\item{screenerFieldsPrefix, screenerFieldsSuffix}{The pre- and suffixes to
pre- and append to the screener identifier in the field names in the
bibliographic database}

\item{prevRoundField}{A field containing information from previous screening
rounds}

\item{duplicateField}{The field containing information about duplicates;
if specified, information from this field is appended to whatever is already
stored in the}

\item{initializeScreenerFields}{Whether to copy the contents of the
initialization field to the screener fields.}

\item{basename}{The base name for the directories and files (appended
with the screener identifiers)}

\item{silent}{Whether to be silent or chatty.}

\item{...}{Any additional arguments are passed to \code{\link[=write_JabRef_Config]{write_JabRef_Config()}}.}
}
\value{
Invisibly, a results object with intermediate results and the
written configuration file
}
\description{
This function prepares a 'screener package': a set of files that one or
more screeners can use to independently screen a set of bibliographic
records.
}
