% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapNewlines.R
\name{zapNewlines}
\alias{zapNewlines}
\title{Zap newlines from a string}
\usage{
zapNewlines(x)
}
\arguments{
\item{x}{The character value or vector}
}
\value{
The resulting character vector
}
\description{
Collapse a character vector into a single value, and replace newlines with
spaces (and replace double spaces with single spaces).
}
\examples{
metabefor::zapNewLines(
  "So here we start,
and here we end"
);
}
