% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_aggregationTree_information.R,
%   R/add_aggregationTree_information_toStudyTree.R
\name{add_aggregationTree_information}
\alias{add_aggregationTree_information}
\alias{add_aggregationTree_information_toStudyTree}
\title{Add information from an aggregation tree to a study tree}
\usage{
add_aggregationTree_information(
  studies,
  aggregationTree,
  fieldName,
  prefixes = NULL,
  suffixes = NULL
)

add_aggregationTree_information_toStudyTree(
  studyTree,
  aggregationTree,
  fieldName,
  prefixes = NULL,
  suffixes = NULL
)
}
\arguments{
\item{aggregationTree}{The aggregation tree}

\item{fieldName}{The field storing the values along which to aggregate}

\item{prefixes, suffixes}{Vector of pre- and suffixes to add to the
\code{fieldname} when storing the values for progressive aggregation}

\item{studyTree}{The study tree}
}
\value{
Invisibly (and, irrelevantly), the study tree
}
\description{
Add information from an aggregation tree to a study tree
}
