% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate_sources.R
\name{check_duplicate_sources}
\alias{check_duplicate_sources}
\title{Check for duplicate sources}
\usage{
check_duplicate_sources(
  primarySources,
  secondarySources = NULL,
  useStringDistances = FALSE,
  stringDistance = 5,
  stringDistanceMethod = "osa",
  charsToZap = "[^A-Za-z0-9]",
  doiCol = "doi",
  matchFully = c("year", "title", "author"),
  matchStart = c(title = 40, author = 30),
  matchEnd = c(title = 40, author = 30),
  forDeduplicationSuffix = "_forDeduplication",
  returnRawStringDistances = FALSE,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{primarySources}{The primary dataframe with sources}

\item{secondarySources}{The secondary dataframe with sources; if supplied
(i.e. for a asymmetric duplicate search), the data frame against which the
primary data frame is checked (i.e. the result specifies, for each entry in
the primary sources, whether it also occurs in the secondary sources).}

\item{useStringDistances}{Whether to use string distances - note that
that can be very slow and take along time if you have thousands of sources.}

\item{stringDistance}{The string distance for titles}

\item{stringDistanceMethod}{Method to use for string distance computation}

\item{charsToZap}{The characters to delete from fields before looking for duplicates}

\item{doiCol}{The name of the column with the DOIs}

\item{matchFully}{A vector of columns to check for full
matches (after 'zapping'). Pass \code{NULL} to not check any columns.}

\item{matchStart, matchEnd}{Named vectors with columns and numbers of
characters to check from the start and from the end. Because requiring full
matches can be too conservative, you can also look at the first or last X
characters. Pass \code{NULL} to not check from the start and from the end, or
pass named vectors where the names are the column names and the elements
are the corresponding numbers of characters to look at for each column. Note:
if a column is also in \code{matchFully}, that takes precedence.}

\item{forDeduplicationSuffix}{Suffix to add to optional deduplication columns}

\item{returnRawStringDistances}{Whether to return the raw string distances
or not (this can be \emph{very} large).}

\item{silent}{Whether to be silent or chatty.}
}
\value{
A vector indicating for each record whether it's a duplicate, with
an attribute called \code{duplicateInfo} that holds more detailed information
and that can be accessed using the \code{\link[=attributes]{attributes()}} function.
}
\description{
This function checks, for two imported dataframes with sources, whether
any of them are duplicates.
}
