% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneRxsProject.R
\name{cloneRxsProject}
\alias{cloneRxsProject}
\title{Clone a full Rxs project object}
\usage{
cloneRxsProject(x, silent = metabefor::opts$get("silent"))
}
\arguments{
\item{x}{The full Rxs project object}

\item{silent}{Whether to be chatty or silent}
}
\value{
The cloned object
}
\description{
Clone a full Rxs project object
}
