% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deinterlace_file.R
\name{deinterlace_file}
\alias{deinterlace_file}
\title{Deinterlace a file}
\usage{
deinterlace_file(x, output = NULL)
}
\arguments{
\item{x}{A charactor vector or a file.}

\item{output}{Optionally, a filename to write the results to.}
}
\value{
The corrected character vector; if \code{output} is not \code{NULL}, that
vector is returned invisibly.
}
\description{
This function removes either \emph{all} odd lines from a file, or \emph{all} even
lines. This can be useful when moving a file between operating systems
(e.g., from Apple to Windows) causes in double line endings.
}
\examples{
metabefor::deinterlace_file(
  (
    "first line",
    "",
    "second (real) line",
    "",
    "third line"
  )
);
}
