% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_sources.R
\name{duplicate_sources}
\alias{duplicate_sources}
\title{Detect duplicate sources}
\usage{
duplicate_sources(
  primarySources,
  secondarySources = NULL,
  stringDistance = 5,
  stringDistanceMethod = "osa",
  titleCharsToZap = "[^A-Za-z0-9]",
  titleCharacterTrimming = 1000,
  doiCol = "doi",
  titleCol = "title",
  forDeduplicationSuffix = "_forDeduplication",
  returnRawStringDistances = FALSE,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{primarySources}{The primary dataframe with sources}

\item{secondarySources}{The secondary dataframe with sources}

\item{stringDistance}{The string distance for titles}

\item{stringDistanceMethod}{Method to use for string distance computation}

\item{titleCharsToZap}{The title characters to delete from titles}

\item{titleCharacterTrimming}{The number of characters to trim titles to}

\item{doiCol}{The name of the column with the DOIs}

\item{titleCol}{The name of the column with the titles}

\item{forDeduplicationSuffix}{Suffix to add to optional deduplication columns}

\item{silent}{}
}
\value{

}
\description{
Detect duplicate sources
}
