% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatMap_from_freqTab.R,
%   R/heatMap_from_studies.R
\name{heatMap_from_freqTab}
\alias{heatMap_from_freqTab}
\alias{heatMap_from_studies}
\title{Produce a heatmap from a frequency table}
\usage{
heatMap_from_freqTab(
  freqTab,
  rowOrder = rownames(freqTab),
  colOrder = colnames(freqTab),
  rowLab = "Row",
  colLab = "Column",
  freqLab = "Frequency",
  plotTitle = paste0("Frequency Table for ", rowLab, " and ", colLab),
  xLabelRotationAngle = 45,
  legend.position = "right",
  fillScale = ggplot2::scale_fill_viridis_c(),
  theme = ggplot2::theme_minimal()
)

heatMap_from_studies(
  x,
  rowRegex,
  colRegex,
  freqTabArgs = NULL,
  rowOrder = rownames(freqTab),
  colOrder = colnames(freqTab),
  rowLab = "Row",
  colLab = "Column",
  freqLab = "Frequency",
  plotTitle = paste0("Frequency Table for ", rowLab, " and ", colLab),
  xLabelRotationAngle = 45,
  legend.position = "right",
  fillScale = ggplot2::scale_fill_viridis_c(),
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{freqTab}{The frequency table (see \code{\link[=studyTreeList_to_freqTab]{studyTreeList_to_freqTab()}})}

\item{rowLab, colLab, freqLab, plotTitle}{The labels to use for the rows (y
axis), columns (x axis), frequencies (fill), and the plot title to use.}

\item{fillScale}{The fill scale to use.}

\item{theme}{The ggplot2 theme to use.}
}
\value{

}
\description{
Produce a heatmap from a frequency table
}
