% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_singleValue.R
\name{get_singleValue_fromTree}
\alias{get_singleValue_fromTree}
\alias{get_singleValue_fromTreeList}
\alias{get_singleValue}
\title{Get a single value from a study tree, list of trees, or studies object}
\usage{
get_singleValue_fromTree(
  x,
  entityId,
  lookInValueLists = TRUE,
  returnDf = FALSE,
  silent = metabefor::opts$get("silent")
)

get_singleValue_fromTreeList(
  x,
  entityId,
  returnDf = TRUE,
  nullValue = 0,
  naValue = NULL,
  warningValues = list(NULL, NA),
  warningFunctions = NULL,
  silent = metabefor::opts$get("silent")
)

get_singleValue(
  x,
  entityId,
  returnDf = TRUE,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{x}{}

\item{entityId}{The entity identifier of the value to get}

\item{lookInValueLists}{Whether to also look inside value lists}

\item{returnDf}{Whether to return a data frame or not}
}
\value{
A list or a dataframe (if \code{returnDf} is \code{TRUE})
}
\description{
Get a single value from a study tree, list of trees, or studies object
}
\examples{

}
