% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valueList_asDf.R
\name{get_valueList_asDf_fromStudyTree}
\alias{get_valueList_asDf_fromStudyTree}
\alias{get_valueList_asDf}
\title{Get a value list as data frame from a study tree, list of trees, or studies object}
\usage{
get_valueList_asDf_fromStudyTree(
  x,
  requiredFields = NULL,
  pathString_regex = NULL,
  flattenVectorsInDf = TRUE,
  silent = metabefor::opts$get("silent")
)

get_valueList_asDf(
  x,
  requiredFields = NULL,
  pathString_regex = NULL,
  flattenVectorsInDf = TRUE,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{x}{The study tree, list of trees, or studies object}

\item{requiredFields}{Fields that have to exist in the target entities
(otherwise, the entity is excluded)}

\item{pathString_regex}{Regex that the target entities' path strings have to
match (otherwise, the entity is excluded)}

\item{flattenVectorsInDf}{When returning a data frame, whether to flatten
vectors into a single character string value, or whether to explode into
multiple rows.}

\item{silent}{Whether to be quiet or chatty.}
}
\value{
A dataframe
}
\description{
Get a value list as data frame from a study tree, list of trees, or studies object
}
\examples{

}
