% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_openalex.R
\name{import_openalex}
\alias{import_openalex}
\title{Import an OpenAlex CSV export}
\usage{
import_openalex(file, extraFields = NULL)
}
\arguments{
\item{file}{The file to import}

\item{extraFields}{Extra fields to store in the resulting dataframe, used
when exporting from CSV}
}
\value{
A data frame.
}
\description{
Import an OpenAlex CSV export
}
\examples{
openAlex_CSV <-
  system.file(
    "extdata",
    "openalex-exports",
    "example-1---CSV-format---works-2024-08-06T12-48-16.csv",
    package = "metabefor"
  );

exampleDat <-
  import_openalex(
    openAlex_CSV
  );

exampleDat$author;
}
