% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_search_results.R
\name{import_search_results}
\alias{import_search_results}
\title{Import search results}
\usage{
import_search_results(
  path,
  dirRegex = ".*",
  fileRegex = "\\\\.ris$",
  dirsToIgnoreRegex = NULL,
  filesToIgnoreRegex = NULL,
  recursive = TRUE,
  perl = TRUE,
  fieldsToCopy = list(doi = c("L3", "DO"), title = "T1"),
  preparatoryReplacements = NULL,
  synthesisr_tag_naming = "best_guess",
  copySep = " || ",
  idFieldName = "original_id",
  parallel = FALSE,
  search_metadataRegex = metabefor::opts$get("search_metadataRegex"),
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{path}{The path to the files with the search results.}

\item{dirRegex}{The regular expression to match subdirectories against.}

\item{fileRegex}{The regular expression to match filenames against.}

\item{dirsToIgnoreRegex, filesToIgnoreRegex}{A regular expression to specify
which directories and files should be ignored.}

\item{recursive}{Whether to recursively read subdirectories.}

\item{perl}{Whether to use Perl regular expressions}

\item{fieldsToCopy}{Fields to copy over to other fields (e.g. \code{L3} sometimes
contains DOIs, instead of \code{DO}). This takes the form of a list of character
vectors, with each vector's name being the new field name (to copy to), and
each character vector listing the fields to copy (to that new field name).
Set to \code{NULL} or pass an empty list to not copy anything over.}

\item{preparatoryReplacements}{Specify any replacements to be made in the
file(s) to import, as a named character vector, where each element's \emph{name}
is the replacement, and the corresponding value is a Perl regular expression
which will (case insensitively) be searched in the file (e.g. to replace
all RIS tags \code{T1} with \code{TI}, pass \code{c("TI" = "^T1")} as value
of \code{preparatoryReplacements}).}

\item{synthesisr_tag_naming}{The value to pass to
the \code{synthesisr::read_refs()} function as argument \code{tag_naming}.}

\item{copySep}{When copying fields over (see \code{fieldsToCopy}), the separator
to use when the new field is not empty (in which case the contents to copy
over will be appended).}

\item{idFieldName}{New name to use for the \code{id} field (column), a reserved
name in JabRef (if \code{NULL}, the field is not renamed).}

\item{parallel}{Whether to use multiple cores for parallel processing.}

\item{search_metadataRegex}{A regular expression to match against the
filenames. If it matches, metadata will be extracted in three capturing
groups, in the order date (using ISO standard 8601 format, i.e. 2022-03-05),
interface, and database, separated by underscores (\verb{_}), with an optional
fourth element, again separated with an underscore, that can be used to
specify which query was run (in case multiple queries are used in the same
database / interface combination and on the same date).}

\item{silent}{Whether to be silent or chatty.}
}
\value{
An object with all the imported information, including, most
importantly, the data frame \code{bibHitDf} with all results.
}
\description{
This is a wrapper around the \code{synthesisr} function \code{read_refs} that
processes the hits in a set of subdirectories in a directory, specifying
the original file, directory, and, if the files were names using a
convention, date, database, and interface (see the \code{search_metadataRegex}
argument).
}
\examples{
### Path to extra files in {metabefor} package
metabefor_files_path <-
  system.file(
    "extdata",
    package = "metabefor"
  ); 

### Path with OpenAlex exports
OpenAlexExport_path <-
  file.path(
    metabefor_files_path,
    "openalex-exports"
  ); 

bibHits_OpenAlex <-
  metabefor::import_search_results(
    OpenAlexExport_path
  );

### Look at the first five titles
bibHits_OpenAlex$bibHitDf$title[1:5]

### Another example, using the filenames to
### provide metadata about the date, database,
### interface, and query specification

### Path with Esbco exports
EbscoExport_path <-
  file.path(
    metabefor_files_path,
    "ebsco-exports"
  ); 
  
bibHits_Ebsco <-
  metabefor::import_search_results(
    EbscoExport_path
  );
  
### Show the databases
metabefor::show_search_hits_by_database(
  bibHits_Ebsco
);
}
