% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_rxsTemplate.R
\name{knit_rxsTemplate}
\alias{knit_rxsTemplate}
\title{Knit the Rxs Template}
\usage{
knit_rxsTemplate(x)
}
\arguments{
\item{x}{The object with the Rxs structure as produced by a call to
\code{\link[=rxs_fromSpecifications]{rxs_fromSpecifications()}}.}
}
\value{
A character vector with class "knit_asis".
}
\description{
This function knits an Rxs template stored in a parsed Rxs specification,
as produced by a call to \code{\link[=rxs_fromSpecifications]{rxs_fromSpecifications()}}, into an
R Markdown (or Quarto) file.
}
\examples{
### Load example Rxs structure
data('rxs_minimal_example_2', package="metabefor");

### Produce the Markdown
knittableMarkdown <-
  knit_rxsTemplate(rxs_minimal_example_2);

### Show the beginning
cat(substr(knittableMarkdown, 1, 2463));
}
