% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knittedFileSansExt.R
\name{knittedFileSansExt}
\alias{knittedFileSansExt}
\title{Return the currently knitted file, without \code{.Rxs.Rmd} extension}
\usage{
knittedFileSansExt(x = NULL)
}
\arguments{
\item{x}{Optionally, a filename. If \code{NULL}, \code{\link[knitr:current_input]{knitr::current_input()}} is
read instead.}
}
\value{
The currently knitted file, without \code{.Rxs.Rmd} extension
}
\description{
Return the currently knitted file, without \code{.Rxs.Rmd} extension
}
\examples{
metabefor::knittedFileSansExt("filename.Rxs.Rmd");
}
