% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_matchingUniqueEntityIdentifiers.R,
%   R/studyTree_matchingUniqueEntityIdentifiers.R
\name{studies_matchingUniqueEntityIdentifiers}
\alias{studies_matchingUniqueEntityIdentifiers}
\alias{studyTree_matchingUniqueEntityIdentifiers}
\title{Look for all entity identifiers that a given pattern}
\usage{
studies_matchingUniqueEntityIdentifiers(
  studies,
  regex,
  includeValueListsOfMatch = TRUE,
  excludeParentWhenValueListReturned = TRUE
)

studyTree_matchingUniqueEntityIdentifiers(
  studyTree,
  regex,
  includeValueListsOfMatch = TRUE,
  excludeParentWhenValueListReturned = TRUE
)
}
\arguments{
\item{studies, studyTree}{The \code{studies} object (i.e. parsed extraction
scripts) or one single \code{studyTree}.}

\item{regex}{The regular expression to use to find entity identifiers}

\item{includeValueListsOfMatch}{Whether to also include the value lists
inside matching entities (useful for quickly selecting e.g. all
results)}

\item{excludeParentWhenValueListReturned}{Whether, if an entity matches,
has a value list as value, and those value lists are returns (i.e.
\code{includeValueListsOfMatch} is \code{TRUE}), the parent entity (that matched
the regular expression) should be excluded.}
}
\value{
A character vector of identifiers
}
\description{
Look for all entity identifiers that a given pattern
}
