% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeName.R
\name{nodeName}
\alias{nodeName}
\title{Check whether a prospective node name is valid, sanitize it, and return it}
\usage{
nodeName(
  x,
  entityName = NULL,
  node = NULL,
  thisEntityText = "this entity (with temporary name '\%s')"
)
}
\arguments{
\item{x}{The prospective node name}

\item{entityName}{The original name of the entity before renaming}

\item{thisEntityText}{Text to insert with the the current name of this
entity, in case it's about to be renamed.}
}
\value{
The (sanitized) node name
}
\description{
Check whether a prospective node name is valid, sanitize it, and return it
}
\examples{
metabefor::nodeName("this_is_valid");
metabefor::nodeName("this is sanitized");
if (FALSE) {
  metabefor::nodeName(
    "_thisThrowsAnError:first character can't be _"
  );
  ### This also throws an error: a reserved name by data.tree
  metabefor::nodeName(
    "parent"
  );
}
}
