% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\docType{data}
\name{opts}
\alias{opts}
\alias{set}
\alias{get}
\alias{reset}
\title{Options for the metabefor package}
\format{
An object of class \code{list} of length 4.
}
\usage{
opts
}
\description{
The \code{metabefor::opts} object contains three functions to set, get, and reset
options used by the metabefor package. Use \code{metabefor::opts$set} to set options,
\code{metabefor::opts$get} to get options, or \code{metabefor::opts$reset} to reset specific or
all options to their default values.
}
\details{
It is normally not necessary to get or set \code{metabefor} options.

The following arguments can be passed:

\describe{
\item{...}{For \code{metabefor::opts$set}, the dots can be used to specify the options
to set, in the format \code{option = value}, for example, \code{utteranceMarker = "\\n"}. For
\code{metabefor::opts$reset}, a list of options to be reset can be passed.}
\item{option}{For \code{metabefor::opts$set}, the name of the option to set.}
\item{default}{For \code{metabefor::opts$get}, the default value to return if the
option has not been manually specified.}
}

The following options can be set:

\describe{
\item{quridPrefix}{The prefix for quasi-unique record identifiers (QURIDs).}

\item{ws}{The worksheet names: a named list with four character values
named \code{entities}, \code{valueTemplates}, \code{definitions}, and \code{instructions}.}
\item{eC}{The entity columns; a named list with character values holding the
names of the columns in the \code{entities} worksheet of the spreadsheet. The
default values are stored in \code{metabefor::opts$get("entityColNames")} - if you
need to override these values, just reproduce that object.}
\item{valueTemplateCols}{The value template columns; a named list with
character values holding the names of the columns in the \code{entities}
worksheet of the spreadsheet. The default values are stored
in \code{metabefor::opts$get("valueTemplateColNames")} - if you need to
override these values, just reproduce that object.}
\item{instructionsCols}{The instructions worksheet columns: a names list with
character values holding the names of the columns in the \code{instructions}
worksheet of the spreadsheet. The default values are stored
in \code{metabefor::opts$get("instructionsColNames")} - if you need to
override these values, just reproduce that object.}
\item{definitionsCols}{The definitions worksheet columns: a names list with
character values holding the names of the columns in the \code{definitions}
worksheet of the spreadsheet. The default values are stored
in \code{metabefor::opts$get("definitionsColNames")} - if you need to
override these values, just reproduce that object.}

\item{indent}{Whether to use indentation to visually organise the Rxs
template. If TRUE, deeper nesting in the Rxs specification's hierarchy
will be visible as deeper indentation.}
\item{indentSpaces}{The number of spaces to use when identing.}
\item{fullWidth}{The maximum width of the Rxs template in characters.}
\item{commentCharacter}{The character used to signify comments - if this is
changed, R will throw errors (unless perhaps it once introduces another
comment symbol).}
\item{fillerCharacter}{The character used after the first character for
filling up space.}
\item{repeatingSuffix}{The suffix to use for the entity identifiers/names
of repeating entities.}

}
}
\examples{
### Get the default Quasi-Unique Record Identifier prefix
metabefor::opts$get(quridPrefix);

### Set it to a custom version, so that every line starts with a pipe
metabefor::opts$set(quridPrefix = "QURID_");

### Check that it worked
metabefor::opts$get(quridPrefix);

### Reset this option to its default value
metabefor::opts$reset(quridPrefix);

### Check that the reset worked, too
metabefor::opts$get(quridPrefix);

}
\keyword{datasets}
