% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rxs.R
\name{rxs_partial}
\alias{rxs_partial}
\alias{knit_print.rxs}
\alias{print.rxs}
\title{Printing a tree of R extraction scripts}
\usage{
rxs_partial(
  x,
  rxsStructure = NULL,
  headingLevel = x$headingLevel,
  printPlots = TRUE,
  echoPartial = FALSE,
  partialFile = NULL,
  ...
)

\method{knit_print}{rxs}(
  x,
  rxsStructure = NULL,
  headingLevel = 3,
  echoPartial = FALSE,
  partialFile = NULL,
  ...
)

\method{print}{rxs}(
  x,
  rxsStructure = NULL,
  headingLevel = 3,
  forceKnitrOutput = FALSE,
  echoPartial = FALSE,
  partialFile = NULL,
  ...
)
}
\arguments{
\item{x}{The study tree}

\item{rxsStructure}{Optionally, the rxs structure}

\item{headingLevel}{The level of the top-most R Markdown heading}

\item{printPlots}{Whether to print plots}

\item{echoPartial}{Whether to echo (show) the R code chunks in the partial}

\item{partialFile}{Optionally (and advanced) the path to a file with an
R Markdown partial.}

\item{...}{Passed on to \code{\link[rmdpartials:partial]{rmdpartials::partial()}}.}
}
\value{
The partial, or for the print function, the printed
information (invisibly).
}
\description{
Printing a tree of R extraction scripts
}
