% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_buildTemplate.R, R/rxs_fromSpecifications.R
\name{rxs_buildTemplate}
\alias{rxs_buildTemplate}
\alias{rxs_fromSpecifications}
\title{Generate an Rxs template}
\usage{
rxs_buildTemplate(
  rxsStructure,
  yamlMetadata = list(title = "Systematic Review Extraction Script Template", author =
    NULL, date = format(Sys.time(), "\%d \%b \%Y at \%H:\%M:\%S")),
  gs_url = NULL,
  indent = TRUE,
  indentSpaces = 2,
  fullWidth = 80,
  commentCharacter = "#",
  fillerCharacter = "#",
  eC = metabefor::opts$get("entityColNames"),
  repeatingSuffix = "__1__",
  silent = FALSE
)

rxs_fromSpecifications(
  gs_url = NULL,
  ws = list(entities = "entities", valueTemplates = "valueTemplates", definitions =
    "definitions", instructions = "instructions"),
  entitiesFilename = NULL,
  valueTemplatesFilename = NULL,
  definitionsFilename = NULL,
  localBackup = list(entities = NULL, valueTemplates = NULL, definitions = NULL),
  outputFile = NULL,
  yamlMetadata = list(title = "Systematic Review Extraction Script Template", author =
    NULL, date = format(Sys.time(), "\%d \%b \%Y at \%H:\%M:\%S")),
  indent = TRUE,
  indentSpaces = 2,
  fullWidth = 80,
  commentCharacter = "#",
  fillerCharacter = "#",
  eC = metabefor::opts$get("entityColNames"),
  valueTemplateCols = metabefor::opts$get("valueTemplateColNames"),
  repeatingSuffix = "__1__",
  rootName = "study",
  silent = metabefor::opts$get("silent"),
  instructionHeadingLevel = 3,
  graphTheme = list(c("fontname", "Arial", "node")),
  returnFullObject = TRUE
)
}
\arguments{
\item{rxsStructure}{The Rxs specification as loaded from a spreadsheet.}

\item{yamlMetadata}{A names list containing character values named
\code{title}, \code{author}, and \code{date} which will be set in the Rmd file that is
the Rxs template as metadata.}

\item{gs_url}{A Google Sheets URL (make sure it's viewable by anybody
with the link!).}

\item{indent}{Whether to use indentation to visually organise the Rxs
template. If TRUE, deeper nesting in the Rxs specification's hierarchy
will be visible as deeper indentation.}

\item{indentSpaces}{The number of spaces to use when identing.}

\item{fullWidth}{The maximum width of the Rxs template in characters.}

\item{commentCharacter}{The character used to signify comments - if this is
changed, R will throw errors (unless perhaps it once introduces another
comment symbol).}

\item{fillerCharacter}{The character used after the first character for
filling up space.}

\item{eC}{The entity columns; a named list with character values holding the
names of the columns in the \code{entities} worksheet of the spreadsheet. The
default values are stored in \code{metabefor::opts$get("entityColNames")} - if you
need to override these values, just reproduce that object.}

\item{repeatingSuffix}{The suffix to use for the entity identifiers/names
of repeating entities.}

\item{silent}{Whether to be silent or chatty.}

\item{ws}{The worksheet names: a named list with four character values
named \code{entities}, \code{valueTemplates}, \code{definitions}, and \code{instructions}.}

\item{entitiesFilename, valueTemplatesFilename, definitionsFilename, }{When
reading from three CSV files, the names of the files containing the
specifications that would otherwise be stored in those three filenames.}

\item{localBackup}{A named list with three character values
named \code{entities}, \code{valueTemplates}, and \code{definitions}, where the information
from the corresponding worksheets will be stored.}

\item{outputFile}{If specified, the path to a file where the Rxs template
will be stored.}

\item{valueTemplateCols}{The value template columns; a named list with
character values holding the names of the columns in the \code{entities}
worksheet of the spreadsheet. The default values are stored
in \code{metabefor::opts$get("valueTemplateColNames")} - if you need to
override these values, just reproduce that object.}

\item{rootName}{The name of the root element}

\item{instructionHeadingLevel}{The top-most heading level for the
instructions.}

\item{graphTheme}{The graph theme to use.}

\item{returnFullObject}{Whether to return the full object or just the
template.}
}
\value{
Either a full object, containing the template and other products,
or just the template.
}
\description{
These functions process an R extraction script template specification in
a spreadsheet format and produce the corresponding R extraction script
template file (an R Markdown file with the extension \code{.rxs.Rmd}).
\code{rxs_fromSpecifications()} can be passed a Google Sheet URL, and uses
\code{rxs_buildTemplate()} in the background to produce the template.
}
