% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_validation.R
\name{rxs_validation}
\alias{rxs_validation}
\title{Title}
\usage{
rxs_validation(studyTree, rxsStructure = NULL)
}
\arguments{
\item{studyTree}{}

\item{rxsStructure}{}
}
\value{

}
\description{
Title
}
