% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_validation.R
\name{rxs_validation}
\alias{rxs_validation}
\title{Validate an Rxs tree}
\usage{
rxs_validation(
  rxsTree,
  stopOnFailure = FALSE,
  stopOnFailure_metadata = stopOnFailure,
  stopOnFailure_entities = stopOnFailure,
  rxsTemplateSpec = NULL,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{rxsTree}{The Rxs tree.}

\item{stopOnFailure}{Whether to throw an error when validation fails.}

\item{stopOnFailure_metadata}{Whether to throw an error when validation
fails, but only if validation fails for the metadata.}

\item{stopOnFailure_entities}{Whether to throw an error when validation
fails, but only if validation fails for the entity values.}

\item{rxsTemplateSpec}{Optionally, the rxs template specification as it is
typically included in rxs templates.}

\item{silent}{Whether to be silent or chatty.}
}
\value{
Invisibly, the Rxs tree (which was altered in place, consistent
with the reference semantics employed by \code{\link[data.tree:Node]{data.tree::Node()}} - so you can
just discard the result).
}
\description{
Validate an Rxs tree
}
