% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_validation.R
\name{rxs_validation}
\alias{rxs_validation}
\title{Validate an Rxs tree}
\usage{
rxs_validation(rxsTree, rxsStructure = NULL)
}
\arguments{
\item{rxsTree}{The Rxs tree}

\item{rxsStructure}{Optionally, the \code{rxsStructure} as resulting from
a call to \code{\link[=rxs_parseSpecifications]{rxs_parseSpecifications()}}.}
}
\value{
Invisibly, the Rxs tree (which was altered in place, consistent
with the reference semantics employed by \code{\link[data.tree:Node]{data.tree::Node()}}.
}
\description{
Validate an Rxs tree
}
