% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_short_doi.R, R/get_short_dois.R
\name{get_short_doi}
\alias{get_short_doi}
\alias{get_short_dois}
\title{Retrieve the short DOI for one or more DOIs}
\usage{
get_short_doi(x = NULL, strip10 = TRUE, throttle = TRUE)

get_short_dois(x = NULL, strip10 = TRUE)
}
\arguments{
\item{x}{A DOI or a vector of DOIs.}

\item{strip10}{Whether to strip the leading \verb{10/}.}

\item{throttle}{Whether to wait for one second if no valid JSON containing
the short doi is returned (\code{throttle=TRUE}) or throw an
error (\code{throttle=FALSE}).}
}
\value{
The short DOI or DOIs.
}
\description{
Retrieve the short DOI for one or more DOIs
}
\examples{
\dontrun{
### Get a short DOI, just the short DOI returned
short_doi(doi = "10.1371/journal.pone.0042793")
short_doi(doi = "10.1890/10-0340.1")
}
\dontrun{
### Get a short DOI, just the short DOI returned
short_doi(doi = "10.1371/journal.pone.0042793")
short_doi(doi = "10.1890/10-0340.1")
}
}
