% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitVectors.R
\name{splitVectors}
\alias{splitVectors}
\title{In a list of vectors, split vectors to multiple list elements}
\usage{
splitVectors(x, sep = "_")
}
\arguments{
\item{x}{The list of vectors}

\item{sep}{The separator to use}
}
\value{
The list with the split vectors
}
\description{
In a list of vectors, split vectors to multiple list elements
}
\examples{
metabefor::splitVectors(
  list(
    c(1, 2, 3),
    4,
    c(5, 6)
  )
);
}
