% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitVectors.R
\name{splitVectors}
\alias{splitVectors}
\title{In a list of vectors, split vectors to multiple list elements}
\usage{
splitVectors(x, sep = "_", fieldname_regex_alwaysFlatten = NULL)
}
\arguments{
\item{x}{The list of vectors}

\item{sep}{The separator to use}

\item{fieldname_regex_alwaysFlatten}{A regular expression that, when it
matches a field name, always causes that field's contents to be flattened.}
}
\value{
The list with the split vectors
}
\description{
In a list of vectors, split vectors to multiple list elements
}
\examples{
metabefor::splitVectors(
  list(
    c(1, 2, 3),
    4,
    c(5, 6)
  )
);

### With list names
metabefor::splitVectors(
  list(
    a = c(1, 2, 3),
    b = 4,
    c = c(5, 6)
  )
);
}
