% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_data_from_list.R,
%   R/supplement_data_from_lists.R, R/supplement_data_with_value.R
\name{supplement_data_from_list}
\alias{supplement_data_from_list}
\alias{supplement_data_from_lists}
\alias{supplement_data_with_value}
\title{Supplement a target entitity node with clustered entity values from another entity node}
\usage{
supplement_data_from_list(
  studyTree,
  targetEntityNodeId,
  sourceEntityNodeIdField_in_targetEntity,
  idField_in_targetEntityNode = NULL,
  idField_in_sourceEntityNode = NULL,
  fieldsToCopy_regex = NULL,
  forceCopyingOfExistingValues = FALSE,
  sourcePathString_regex = NULL,
  targetPathString_regex = NULL,
  prefix = "",
  suffix = "",
  silent = metabefor::opts$get("silent")
)

supplement_data_from_lists(
  studyTree,
  sourceEntityNodeIdField_in_targetEntity,
  idField_in_targetEntityNode = NULL,
  idField_in_sourceEntityNode = NULL,
  fieldsToCopy_regex = NULL,
  sourcePathString_regex = NULL,
  targetPathString_regex = NULL,
  forceCopyingOfExistingValues = FALSE,
  prefix = "",
  suffix = "",
  silent = metabefor::opts$get("silent")
)

supplement_data_with_value(
  studyTree,
  targetEntityNodeId,
  sourceEntityNodeIdField_in_targetEntity,
  idField_in_targetEntityNode = NULL,
  idField_in_sourceEntityNode = NULL,
  fieldsToCopy_regex = NULL,
  forceCopyingOfExistingValues = FALSE,
  sourcePathString_regex = NULL,
  targetPathString_regex = NULL,
  prefix = "",
  suffix = "",
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{studyTree}{The study tree}

\item{targetEntityNodeId}{The identifier of the target entity node (the node to
supplement)}

\item{sourceEntityNodeIdField_in_targetEntity}{Inside the target entity node,
the name holding the identifier of the source entity node (the node supplying
the data). Note that that field is itself an entity as specified in the entity
specification spreadsheet.}

\item{idField_in_targetEntityNode}{\code{NULL} if the target node's name is also its
identifier; otherwise, the name of the field containing the identifier in
the target node.}

\item{idField_in_sourceEntityNode}{\code{NULL} if the source node's name is also its
identifier; otherwise, the name of the field containing the identifier in
the source node.}

\item{fieldsToCopy_regex}{A regular expression that can optionally be
used to select fields to copy over from the source node to the target node.}

\item{sourcePathString_regex, targetPathString_regex}{Regular expressions
that must match the path string of the source of target node.}

\item{prefix, suffix}{A text string to prepend and append to the names of the
values that are copied.}

\item{silent}{Whether to be quiet or chatty.}

\item{overwriteExistingValues}{Whether to overwrite existing values if
those are encountered in the target entity.}
}
\value{



}
\description{
These function supplement a entity nodes with clustered entity values
from other entities specified in the target entity nodes. Both the
target entity and the source entity have to be clustering entities.
Use \code{supplement_data_from_list()} for one node, and
\code{supplement_data_from_list()} for all nodes containing the field that
identifies the node to copy the values from (i.e. as specified in
\code{sourceEntityNodeIdField_in_targetEntity}).

This function supplements an entity node with a value from another entity
node.
}
\details{
Both the
target entity and the source entity have to be clustering entities.
Use \code{supplement_data_from_list()} for one node, and
\code{supplement_data_from_list()} for all nodes containing the field that
identifies the node to copy the values from (i.e. as specified in
\code{sourceEntityNodeIdField_in_targetEntity}).
}
