% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_studyTree_from_txs.R
\name{supplement_studyTrees_from_txs}
\alias{supplement_studyTrees_from_txs}
\title{Supplement a parsed studies object with entities from a tabulated extraction spreadsheet (txs) specification}
\usage{
supplement_studyTrees_from_txs(
  studies,
  txs_specs,
  stopOnErrors = FALSE,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{studies}{The parsed studies object, containing a list of study trees.}

\item{txs_specs}{The txs specifications, as a link to a publicly readable
Google spreadsheet, an Excel file (if you have {\code{openxlsx}} installed),
an SPSS dataset (if you have {\code{haven}} installed), or a comma separated
values file (which will be read with \code{\link[=read.csv]{read.csv()}}).}

\item{stopOnErrors}{Whether to throw an error or show a warning (or just
use \code{cat} to show a message is \code{silent=FALSE}) when encountering errors.}

\item{silent}{Whether to the chatty or silent.}
}
\value{
Invisibly, the studies object. Note that the study trees will be
changed in place given \code{data.tree}'s pass-by-reference logic; so you can
discard the result.
}
\description{
Sometimes, you forget to extract one or more entities. In such cases, it can
be more efficient to create a spreadsheet with the omitted entities and then
add those into the existing studies object. For that purpose, the tabulated
extraction spreadsheet (txs) format exists. This is a rectangular table
with the following columns: \code{study_identification_entity_id},
\code{study_identification_value}, \code{parent_entity_id}, \code{entity_id}, and \code{value}.
}
\details{
If \code{study_identification_entity_id} is empty (i.e. does not contain any
non-whitespace characters), \code{study_identification_value} will be assumed to
be the name of the relevant study tree in the studies object.
}
