% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_rxsTree.R, R/tidy_rxsTrees.R
\name{tidy_rxsTree}
\alias{tidy_rxsTree}
\alias{tidy_rxsTrees}
\title{Flatten one or more Rxs trees into a tidy data frame}
\usage{
tidy_rxsTree(
  rxsTree,
  sourceId = "",
  flattenToString = TRUE,
  silent = metabefor::opts$get("silent")
)

tidy_rxsTrees(
  x,
  flattenToString = TRUE,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{rxsTree}{The Rxs tree to convert}

\item{sourceId}{The source identifier}

\item{flattenToString}{Whether to flatten lists to a string
or explode to separate variables.}

\item{silent}{Whether to be silent or chatty.}

\item{x}{A list of Rxs trees}
}
\value{
A data frame
}
\description{
Flatten one or more Rxs trees into a tidy data frame
}
