% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_studyTree.R, R/tidy_studyTrees.R
\name{tidy_studyTree}
\alias{tidy_studyTree}
\alias{tidy_studyTrees}
\title{Flatten one or more study trees into a tidy data frame}
\usage{
tidy_studyTree(
  studyTree,
  studyName = "",
  eC = metabefor::opts$get("entityColNames"),
  flattenToString = TRUE,
  silent = metabefor::opts$get("silent")
)

tidy_studyTrees(
  x,
  eC = metabefor::opts$get("entityColNames"),
  flattenToString = TRUE,
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{studyTree}{The study tree to convert}

\item{studyName}{The name of the study}

\item{eC}{The names of the entity columns}

\item{flattenToString}{Whether to flatten lists to a string
or explode to separate variables.}

\item{x}{A list of study trees}
}
\value{
A data frame
}
\description{
Flatten one or more study trees into a tidy data frame
}
