% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_vectorValue_to_valueList.R,
%   R/studyTreeList_vectorValue_to_valueList.R,
%   R/studyTree_vectorValue_to_valueList.R
\name{studies_vectorValue_to_valueList}
\alias{studies_vectorValue_to_valueList}
\alias{studyTreeList_vectorValue_to_valueList}
\alias{studyTree_vectorValue_to_valueList}
\title{In an entity, convert a vector to a value list}
\usage{
studies_vectorValue_to_valueList(x, entityId, allPossibleValues = NULL)

studyTreeList_vectorValue_to_valueList(x, entityId, allPossibleValues = NULL)

studyTree_vectorValue_to_valueList(x, entityId, allPossibleValues = NULL)
}
\arguments{
\item{x}{A study tree}

\item{entityId}{The entity Id}

\item{allPossibleValues}{A list of all possible values; if provided,
the value list will contain all those.}
}
\value{

}
\description{
For example, \code{c('red', 'brown')} becomes \code{list(red = 1, brown = 1)}.
}
