% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_vectorValue_to_valueList.R,
%   R/studyTreeList_vectorValue_to_valueList.R,
%   R/studyTree_vectorValue_to_valueList.R, R/vectorValue_to_valueList.R
\name{studies_vectorValue_to_valueList}
\alias{studies_vectorValue_to_valueList}
\alias{studyTreeList_vectorValue_to_valueList}
\alias{studyTree_vectorValue_to_valueList}
\alias{vectorValue_to_valueList}
\title{In an entity, convert a vector to a value list}
\usage{
studies_vectorValue_to_valueList(x, entityId, allPossibleValues = NULL)

studyTreeList_vectorValue_to_valueList(x, entityId, allPossibleValues = NULL)

studyTree_vectorValue_to_valueList(x, entityId, allPossibleValues = NULL)

vectorValue_to_valueList(
  value,
  allPossibleValues = NULL,
  labels = NULL,
  valueForPresence = 1,
  valueForAbsence = 0
)
}
\arguments{
\item{x}{A study tree}

\item{entityId}{The entity Id}

\item{allPossibleValues}{A list of all possible values}

\item{value}{The vector value to convert}

\item{labels}{Optionally, a named vector with labels to replace the actual
values with; names are values, values are labels.}

\item{valueForPresence}{Value to assign for the elements in the vector}

\item{valueForAbsence}{Value to assign for the elements
in \code{allPossibleValues} that do not occur in the vector}
}
\value{


The list
}
\description{
For example, \code{c('red', 'brown')} becomes \code{list(red = 1, brown = 1)}.
}
