% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_JabRef_Config.r
\name{write_JabRef_Config}
\alias{write_JabRef_Config}
\title{Write a JabRef configuration file}
\usage{
write_JabRef_Config(
  outputPath,
  screeners = c("a", "b"),
  screenerFieldsPrefix = "screener",
  screenerFieldsSuffix = "status",
  screenerConfidencePrefix = "screener",
  screenerConfidenceSuffix = "confidence",
  fields = c("title", "abstract"),
  duplicateField = NULL,
  sortField = "title",
  sortDesc = FALSE,
  screeningType = "screening",
  jabrefVersion = "2.11.1",
  jabrefXML = NULL,
  batFile = NULL,
  generateSettingRemovalBatchFile = FALSE,
  screenerSuffixInFilename = FALSE
)
}
\arguments{
\item{outputPath}{The path to write the file(s) to.}

\item{screeners}{The strings to use to identify the screeners.}

\item{screenerFieldsPrefix, screenerFieldsSuffix}{The prefix and suffix to
attach to the screener identification strings to create the fields where
the screeners store their decision.}

\item{screenerConfidencePrefix, screenerConfidenceSuffix}{The prefix and
suffix to attach to the screener identification strings to create the
fields where the screeners store their confidence in their decision.}

\item{fields}{The fields that have to be screened.}

\item{duplicateField}{The field}

\item{sortField}{The field to sort entries by.}

\item{sortDesc}{Whether to sort in descending or ascending order.}

\item{screeningType}{The type of screening (added to the filenames).}

\item{jabrefVersion}{The version of JabRef to format the configuration file
for.}

\item{jabrefXML}{Optionally, the path to a file with a jabrefXML 'skeleton',
in which certain, as yet undocumented, replacements will be made (but you
can check the source code of the function in the meantime).}

\item{batFile}{Optionally, contents of the batch file to write.}

\item{generateSettingRemovalBatchFile}{Whether to write a batch file that
edits the Windows registry to remove JabRef configuration, which may
sometimes be necessary to get JabRef to use the XML configuration file.}

\item{screenerSuffixInFilename}{Whether to add the screener identifier as
a suffix to the file that is written.}
}
\value{
Invisibly, an object with the configuration file and intermediate
steps.
}
\description{
This function writes one or more JabRef configuration files that set
a number of fields to be visible to screeners, so that you can mask the
screeners from fields like journal, authors, and publication year.
}
\examples{
### Get the path to a temporary directory
tempPath <- tempdir();

### Write screener packages
metabefor::write_JabRef_Config(
  tempPath,
  screeners = "A"
);

### Load written file
jabRefConfig <-
  readLines(
    file.path(
      tempPath,
      "jabref.xml"
    )
  );

### Show beginning of file
cat(jabRefConfig[1:10], sep="\n");

### Show bit in the middle with specific screening info
cat(jabRefConfig[39:41], sep="\n");
}
